### 2021-11-19 #3924 @whedon
[PRE REVIEW]: GXBeam: A Pure Julia Implementation of Geometrically Exact Beam Theory
**Submitting author:** @taylormcd (<a href="http://orcid.org/0000-0001-7993-5703">Taylor McDonnell</a>)
**Repository:** <a href="https://github.com/byuflowlab/GXBeam.jl" target ="_blank">https://github.com/byuflowlab/GXBeam.jl</a>
**Version:** v0.3.1
