### 2021-12-02 896e1b1
fix point mass test comparison
