### 2021-12-16 #3924 @Andlon
Hi @prashjha: I appreciate the recommendation from @Kevin-Mattheus-Moerman, but I unfortunately don't have enough experience with Julia nor beam theory to be comfortable about this!
