+++ test/runtests.jl
@@ -1479,14 +1479,13 @@ end
 
 @testset "Point Masses" begin
 
-    nelem = 5
     nodes = [[0,i,0] for i in 0:.1:1]
     nelem = length(nodes)-1
     start = 1:nelem
     stop =  2:(nelem+1)
 
     frames = fill(wiener_milenkovic(rand(3)), nelem)
-    stiffness = fill(Symmetric(rand(6,6)), nelem)
+    compliance = fill(Symmetric(rand(6,6)), nelem)
     mass = fill(Symmetric(rand(6,6)), nelem)
        
     prescribed_conditions = Dict(1 => PrescribedConditions(ux=0, uy=0, uz=0, theta_x=0, theta_y=0, theta_z=0));
@@ -1494,7 +1493,7 @@ end
     # assembly of mass-containing beam elements
 
     assembly = GXBeam.Assembly(nodes, start, stop;
-        stiffness=stiffness, 
+        compliance=compliance, 
         frames=frames, 
         mass=mass);
     
@@ -1509,7 +1508,7 @@ end
     # assembly of massless beam elements with point masses
     
     assembly = GXBeam.Assembly(nodes, start, stop;
-        stiffness=stiffness, 
+        compliance=compliance, 
         frames=frames);
    
     point_masses = Dict{Int, PointMass{Float64}}()
