### 2021-12-18 ed25219
fix default force scaling corner case
