### 2021-12-18 0bd6a47
CompatHelper: add new compat entry for SciMLBase at version 1, (keep existing compat)
