+++ Project.toml
@@ -23,6 +23,7 @@ LinearMaps = "2, 3"
 NLsolve = "4"
 OffsetArrays = "1"
 Roots = "1"
+SciMLBase = "1"
 StaticArrays = "0.12, 1.0"
 WriteVTK = "1"
 julia = "1.5"
