### 2022-01-19 a4bc9c9
add performance tests
