+++ paper/paper.bib
@@ -1,8 +1,10 @@
+% Encoding: UTF-8
 @Article{Bezanson2017,
   author    = {Bezanson, Jeff and Edelman, Alan and Karpinski, Stefan and Shah, Viral B},
   title     = {Julia: A fresh approach to numerical computing},
   number    = {1},
   pages     = {65--98},
+  doi       = {10.1137/141000671},
   url       = {https://doi.org/10.1137/141000671},
   volume    = {59},
   journal   = {SIAM review},
@@ -124,6 +126,55 @@
   comment      = {doi:10.2514/5.9781600866821.0059.0102},
   issn         = {978-1-56347-697-6},
   journaltitle = {Nonlinear Composite Beam Theory},
+}
+
+@Article{Betsch2002,
+  author    = {Betsch, P. and Steinmann, P.},
+  title     = {Frame-indifferent beam finite elements based upon the geometrically exact beam theory},
+  journal   = {Int. J. Numer. Meth. Engng.},
+  year      = {2002},
+  volume    = {54},
+  number    = {12},
+  month     = aug,
+  pages     = {1775--1788},
+  issn      = {0029-5981},
+  doi       = {10.1002/nme.487},
+  url       = {https://doi.org/10.1002/nme.487},
+  abstract  = {Abstract This paper describes a new beam finite element formulation based upon the geometrically exact beam theory. In contrast to many previously proposed beam finite element formulations the present discretization approach retains the frame-indifference (or objectivity) of the underlying beam theory. The space interpolation of rotational degrees-of-freedom is circumvented by the introduction of a reparameterization of the weak form corresponding to the equations of motion of the geometrically exact beam theory. Copyright ? 2002 John Wiley & Sons, Ltd.},
+  comment   = {https://doi.org/10.1002/nme.487},
+  keywords  = {non-linear beam elements, finite rotations, constraints},
+  publisher = {John Wiley \& Sons, Ltd},
+}
+
+@Article{Auricchio2008,
+  author        = {Auricchio, F. and Carotenuto, P. and Reali, A.},
+  title         = {On the geometrically exact beam model: A consistent, effective and simple derivation from three-dimensional finite-elasticity},
+  journal       = {International Journal of Solids and Structures},
+  year          = {2008},
+  volume        = {45},
+  number        = {17},
+  month         = aug,
+  pages         = {4766--4781},
+  issn          = {0020-7683},
+  url           = {https://www.sciencedirect.com/science/article/pii/S0020768308001790},
+  __markedentry = {[taylor:]},
+  abstract      = {The goal of the present work is to emphasize some intrinsic features of the three-dimensional beam model originally proposed by Simo [Simo, J.C., 1985. A finite-strain beam formulation. The three-dimensional dynamic problem. Part I. Comput. Methods Appl. Mech. Eng. 49, 55-70] and to derive the model equations for the general finite-deformation case as well as for the finite-deformation small-strain case in a consistent but simpler way with respect to what it is generally done in the literature.},
+  keywords      = {Beam, Finite-strain, Geometrically exact, Extended polar decomposition, Small-strain, Virtual work},
+}
+
+@Article{Ibrahimbegovic1995,
+  author        = {Ibrahimbegović, Adnan},
+  title         = {On finite element implementation of geometrically nonlinear Reissner's beam theory: three-dimensional curved beam elements},
+  journal       = {Computer Methods in Applied Mechanics and Engineering},
+  year          = {1995},
+  volume        = {122},
+  number        = {1},
+  month         = apr,
+  pages         = {11--26},
+  issn          = {0045-7825},
+  url           = {https://www.sciencedirect.com/science/article/pii/004578259500724F},
+  __markedentry = {[taylor:6]},
[…]
