### 2022-01-20 5e718fa
add convergence plot to examples
