### 2022-01-20 654d63f
add note about convergence
