### 2022-01-20 10eab9f
adjust wording in limitations section
