### 2022-01-20 291d08b
remove labels from convergence plots
