+++ paper/paper.bib
@@ -112,6 +112,105 @@
   year    = {2017},
 }
 
+@article{Ritto2002,
+  title={On the differentiation of the Rodrigues formula and its significance for the vector-like parameterization of Reissner--Simo beam theory},
+  author={Ritto-Corr{\^e}a, M and Camotim, D19321661033},
+  journal={International Journal for Numerical Methods in Engineering},
+  volume={55},
+  number={9},
+  pages={1005--1032},
+  year={2002},
+  publisher={Wiley Online Library}
+}
+
+@article{Betsch2002,
+  title={Frame-indifferent beam finite elements based upon the geometrically exact beam theory},
+  author={Betsch, P and Steinmann, P},
+  journal={International journal for numerical methods in engineering},
+  volume={54},
+  number={12},
+  pages={1775--1788},
+  year={2002},
+  publisher={Wiley Online Library}
+}
+
+@article{Ibrahimbegovic1998,
+  title={Finite rotations in dynamics of beams and implicit time-stepping schemes},
+  author={Ibrahimbegovi{\'c}, Adnan and Mikdad, Mazen Al},
+  journal={International Journal for Numerical Methods in Engineering},
+  volume={41},
+  number={5},
+  pages={781--814},
+  year={1998},
+  publisher={Wiley Online Library}
+}
+
+@article{Jelenic1995,
+  title={A kinematically exact space finite strain beam model—finite element formulation by generalized virtual work principle},
+  author={Jeleni{\'c}, Gordan and Saje, Miran},
+  journal={Computer Methods in Applied Mechanics and Engineering},
+  volume={120},
+  number={1-2},
+  pages={131--161},
+  year={1995},
+  publisher={Elsevier}
+}
+
+@article{Ibrahimbegovic1995a,
+  title={Computational aspects of vector-like parametrization of three-dimensional finite rotations},
+  author={Ibrahimbegovi{\'c}, Adnan and Frey, Francois and Ko{\v{z}}ar, Ivica},
+  journal={International Journal for Numerical Methods in Engineering},
+  volume={38},
+  number={21},
+  pages={3653--3673},
+  year={1995},
+  publisher={Wiley Online Library}
+}
+
+@article{Ibrahimbegovic1995b,
+  title={On finite element implementation of geometrically nonlinear Reissner's beam theory: three-dimensional curved beam elements},
+  author={Ibrahimbegovi{\'c}, Adnan},
+  journal={Computer methods in applied mechanics and engineering},
+  volume={122},
+  number={1-2},
+  pages={11--26},
+  year={1995},
+  publisher={Elsevier}
+}
+
+@article{Cardona1988,
+  title={A beam finite element non-linear theory with finite rotations},
+  author={Cardona, A and Geradin, M},
+  journal={International journal for numerical methods in engineering},
+  volume={26},
+  number={11},
+  pages={2403--2438},
+  year={1988},
+  publisher={Wiley Online Library}
+}
+
+@article{Iura1988,
+  title={Dynamic analysis of finitely stretched and rotated three-dimensional space-curved beams},
+  author={Iura, M and Atluri, SN},
+  journal={Computers \& Structures},
+  volume={29},
+  number={5},
+  pages={875--889},
+  year={1988},
+  publisher={Elsevier}
+}
+
+@article{Dvorkin1988,
+  title={On a non-linear formulation for curved Timoshenko beam elements considering large displacement/rotation increments},
+  author={Dvorkin, Eduardo N and Onte, Eugenio and Oliver, Javier},
+  journal={International Journal for Numerical Methods in Engineering},
+  volume={26},
+  number={7},
+  pages={1597--1613},
+  year={1988},
[…]
