+++ paper/paper.md
@@ -21,46 +21,18 @@ bibliography: paper.bib
 
 # Summary
 
-When the cross sections of a three-dimensional structure are small compared to the length of the structure, beam theory may be used to efficiently model the structure's three-dimensional behavior.  Applications of beam theory include, but are not limited to, the structural modeling of buildings, bridges, aircraft, helicopter blades, and wind turbines.  When deflections are small, linear beam theories may be used to model the behavior of slender structures.  When deflections become significant, such as encountered when modeling high aspect ratio wings and/or large wind turbine blades, nonlinearities associated with geometric deformations must be accounted for.  
+When the cross sections of a three-dimensional structure are small compared to the length of the structure, beam theory may be used to efficiently model the structure's three-dimensional behavior.  Applications of beam theory include, but are not limited to, the structural modeling of buildings, bridges, aircraft, helicopter blades, and wind turbines.  When deflections are small, linear beam theories may be used to model the behavior of slender structures.  When deflections become significant, such as encountered when modeling high aspect ratio wings or large wind turbine blades, nonlinearities associated with geometric deformations must be accounted for.  
 
-Geometrically exact beam theory, as pioneered by Reissner [@Reissner1973], captures all of the nonlinearities associated with large deflections and/or rotations, assuming strains are small.  This beam theory was extended to model general three dimensional dynamics by Simo [@Simo1985] and Simo and Vu-Quoc [@Simo1986].  Betsch and Steinmann later reparameterized the equations of motion of geometrically exact beam theory in order to preserve the geometrical exactness of the theory.
-
-  Since it was introduced, this beam theory has been extended and used by many researchers, with significant efforts directed towards modeling finite rotations.  These improvements have developed geometrically exact beam theory to the point where it is now the de-facto standard for the modeling of highly flexible beams.
+Geometrically exact beam theory, as pioneered by Reissner [@Reissner1973], captures all of the nonlinearities associated with large deflections and rotations, assuming strains are small.  This beam theory was extended to model general three dimensional dynamics by Simo [@Simo1985] and Simo and Vu-Quoc [@Simo1986;@Simo1988] and has since been the extended and used by many researchers[@Kondoh1986;@Dvorkin1988;@Iura1988;@Cardona1988;@Cardona1991;@Ibrahimbegovic1995a;@Ritto2002;@Ibrahimbegovic1995b;@Jelenic1995;@Ibrahimbegovic1998;@Crisfield1998;@Jelenic1998;@Jelenic1999;@Betsch2002@Kapani2003;].  The various improvements to geometrically exact beam theory proposed by researchers throughout the years have progressed geometrically exact beam theory to the point where it has now become an invaluable resource for analyzing and modeling highly flexible slender structures.
 
 # Statement of Need
 
 `GXBeam` is a geometrically exact beam theory package which is written completely in the Julia programming language [@Bezanson2017].  It was originally based on the open source code `GEBT` and its associated papers [@Yu2012;@Wang2017], which adopt the mixed formulation of geometrically exact beam theory presented by Hodges[@Hodges].  When combined with a beam cross sectional analysis, such as a variational asymptotic beam sectional analysis [@Hodges2006a], this geometrically exact beam theory formulation constitutes an efficient and accurate replacement for a full three-dimensional structural analysis.
 
[…]
