+++ paper/paper.bib
@@ -20,6 +20,34 @@
   year    = {2016},
 }
 
+@article{Simo1986,
+  title={On the dynamics of flexible beams under large overall motions—The plane case: Part II},
+  author={Simo, J Cl and Vu-Quoc, L},
+  year={1986}
+}
+
+@article{Simo1988,
+  title={On the dynamics in space of rods undergoing large motions—a geometrically exact approach},
+  author={Simo, Juan Carlos and Vu-Quoc, Loc},
+  journal={Computer methods in applied mechanics and engineering},
+  volume={66},
+  number={2},
+  pages={125--161},
+  year={1988},
+  publisher={Elsevier}
+}
+
+@article{Simo1985,
+  title={A finite strain beam formulation. The three-dimensional dynamic problem. Part I},
+  author={Simo, Juan C},
+  journal={Computer methods in applied mechanics and engineering},
+  volume={49},
+  number={1},
+  pages={55--70},
+  year={1985},
+  publisher={Elsevier}
+}
+
 @Software{Mogensen2020,
   author    = {Patrick Kofod Mogensen and Kristoffer Carlsson and Sébastien Villemot and Spencer Lyon and Matthieu Gomez and Christopher Rackauckas and Tim Holy and David Widmann and Tony Kelman and Daniel Karrasch},
   title     = {NLsolve.jl},
