+++ paper/paper.bib
@@ -85,16 +85,14 @@
  month = may,
 }
 
-@techreport{Mogensen2020,
- author = {Myers, Clayton and Gomez, Matthew and Lamppa, Derek and Webb, Timothy and Yager-Elorriaga, David and Hutsel, Brian and Jennings, Christopher and Knapp, Patrick and Kossow, Michael and Lucero, Larry and Obregon, Robert and Steiner, Adam and Sinars, Daniel},
- title = {Developing inductively driven diagnostic X-ray sources to enable transformative radiography and diffraction capabilities on Z.},
- doi = {10.2172/1747000},
- url = {https://doi.org/10.2172/1747000},
- month = dec,
- publisher = {Zenodo},
- year = {2020},
- source = {Crossref},
- institution = {Office of Scientific and Technical Information (OSTI)},
+@Software{Mogensen2020,
+  author    = {Patrick Kofod Mogensen and Kristoffer Carlsson and Sébastien Villemot and Spencer Lyon and Matthieu Gomez and Christopher Rackauckas and Tim Holy and David Widmann and Tony Kelman and Daniel Karrasch},
+  title     = {NLsolve.jl},
+  doi       = {10.5281/zenodo.4404703},
+  url       = {https://doi.org/10.5281/zenodo.4404703},
+  month     = dec,
+  publisher = {Zenodo},
+  year      = {2020},
 }
 
 @software{Polanco2021,
