### 2022-01-20 8aa0b9d
update list of references
