+++ paper/paper.md
@@ -23,7 +23,7 @@ bibliography: paper.bib
 
 When the cross sections of a three-dimensional structure are small compared to the length of the structure, beam theory may be used to efficiently model the structure's three-dimensional behavior.  Applications of beam theory include, but are not limited to, the structural modeling of buildings, bridges, aircraft, helicopter blades, and wind turbines.  When deflections are small, linear beam theories may be used to model the behavior of slender structures.  When deflections become significant, such as encountered when modeling high aspect ratio wings or large wind turbine blades, nonlinearities associated with geometric deformations must be accounted for.  
 
-Geometrically exact beam theory, as pioneered by Reissner [@Reissner1973], captures all of the nonlinearities associated with large deflections and rotations, assuming strains are small.  This beam theory was extended to model general three dimensional dynamics by Simo [@Simo1985] and Simo and Vu-Quoc [@Simo1986;@Simo1988] and has since been the extended and used by many researchers[@Dvorkin1988,@Cardona1988,@Iura1988,@Jelenic1995,Ibrahimbegovic1995,@Ibrahimbegovic1995a,@Ibrahimbegovic1998,@Ritto2002,@Betsch2002].  The various improvements to geometrically exact beam theory proposed by researchers throughout the years have progressed geometrically exact beam theory to the point where it has now become an invaluable resource for analyzing and modeling highly flexible slender structures.
+Geometrically exact beam theory, as pioneered by Reissner [@Reissner1973], captures all of the nonlinearities associated with large deflections and rotations, assuming strains are small.  This beam theory was extended to model general three dimensional dynamics by Simo [@Simo1985] and Simo and Vu-Quoc [@Simo1986;@Simo1988] and has since been the extended and used by many researchers[@Dvorkin1988;@Cardona1988;@Iura1988;@Jelenic1995;@Ibrahimbegovic1995;@Ibrahimbegovic1995a;@Ibrahimbegovic1998;@Ritto2002;@Betsch2002].  The various improvements to geometrically exact beam theory proposed by researchers throughout the years have progressed geometrically exact beam theory to the point where it has now become an invaluable resource for analyzing and modeling highly flexible slender structures.
 
 # Statement of Need
 
