+++ paper/paper.bib
@@ -15,7 +15,7 @@
 }
 
 @article{Revels2016,
- author = {\{Revels\}, J. and \{Lubin\}, M. and \{Papamarkou\}, T.},
+ author = {{Revels}, J. and {Lubin}, M. and {Papamarkou}, T.},
  title = {Forward-Mode Automatic Differentiation in {{J}ulia}},
  url = {https://arxiv.org/abs/1607.07892},
  journal = {arXiv:1607.07892 [cs.MS]},
@@ -124,7 +124,7 @@
 
 @incollection{Hodges2006,
  author = {Hodges, Dewey H},
- booktitle = {\{Nonlinear Composite Beam Theory\}},
+ booktitle = {{Nonlinear Composite Beam Theory}},
  date = {2006-04},
  title = {{One-Dimensional Theory of Beams}},
  doi = {10.2514/5.9781600866821.0103.0142},
@@ -204,7 +204,6 @@
  doi = {10.1002/nme.532},
  source = {Crossref},
  url = {https://doi.org/10.1002/nme.532},
- subtitle = {DIFFERENTIATION OF RODRIGUES FORMULA},
  issn = {0029-5981},
  month = aug,
 }
@@ -255,22 +254,6 @@
  url = {https://doi.org/10.1002/nme.1620382107},
  issn = {0029-5981, 1097-0207},
  month = nov,
-}
-
-@article{Ibrahimbegovic1995b,
- author = {Ibrahimbegovi\'c, Adnan},
- title = {On finite element implementation of geometrically nonlinear Reissner's beam theory: {Three-dimensional} curved beam elements},
- journal = {Comput. Method. Appl. M.},
- volume = {122},
- number = {1-2},
- pages = {11--26},
- year = {1995},
- publisher = {Elsevier BV},
- doi = {10.1016/0045-7825(95)00724-f},
- source = {Crossref},
- url = {https://doi.org/10.1016/0045-7825(95)00724-f},
- issn = {0045-7825},
- month = apr,
 }
 
 @article{Cardona1988,
@@ -354,25 +337,6 @@
  keywords = {non-linear beam elements, finite rotations, constraints},
  publisher = {Wiley},
  source = {Crossref},
-}
-
-@article{Auricchio2008,
- author = {Auricchio, F. and Carotenuto, P. and Reali, A.},
- title = {On the geometrically exact beam model: {A} consistent, effective and simple derivation from three-dimensional finite-elasticity},
- issn = {0020-7683},
- number = {17},
- pages = {4766--4781},
- url = {https://doi.org/10.1016/j.ijsolstr.2008.04.015},
- volume = {45},
- abstract = {The goal of the present work is to emphasize some intrinsic features of the three-dimensional beam model originally proposed by Simo [Simo, J.C., 1985. A finite-strain beam formulation. The three-dimensional dynamic problem. Part I. Comput. Methods Appl. Mech. Eng. 49, 55-70] and to derive the model equations for the general finite-deformation case as well as for the finite-deformation small-strain case in a consistent but simpler way with respect to what it is generally done in the literature.},
- groups = {[taylor:]},
- journal = {Int. J. Solids Struct.},
- keywords = {Beam, Finite-strain, Geometrically exact, Extended polar decomposition, Small-strain, Virtual work},
- month = aug,
- year = {2008},
- doi = {10.1016/j.ijsolstr.2008.04.015},
- source = {Crossref},
- publisher = {Elsevier BV},
 }
 
 @article{Ibrahimbegovic1995,
