+++ paper/paper.md
@@ -27,7 +27,7 @@ Geometrically exact beam theory, as pioneered by Reissner [@Reissner1973], captu
 
 # Statement of Need
 
-`GXBeam` is a geometrically exact beam theory package which is written completely in the Julia programming language [@Bezanson2017].  It was originally based on the open source code `GEBT` and its associated papers [@Yu2012;@Wang2017], which adopt the mixed formulation of geometrically exact beam theory presented by Hodges[@Hodges1990].  When combined with a beam cross sectional analysis, such as a variational asymptotic beam sectional analysis [@Hodges2006a], this geometrically exact beam theory formulation constitutes an efficient and accurate replacement for a full three-dimensional structural analysis.
+`GXBeam` is a geometrically exact beam theory package which is written completely in the Julia programming language [@Bezanson2017].  It was originally based on the open source code `GEBT` and its associated papers [@Yu2012;@Wang2017], which adopt the mixed formulation of geometrically exact beam theory presented by Hodges [@Hodges1990].  When combined with a beam cross sectional analysis, such as a variational asymptotic beam sectional analysis [@Hodges2006a], this geometrically exact beam theory formulation constitutes an efficient and accurate replacement for a full three-dimensional structural analysis.
 
 One of the key advantages of `GXBeam` relative to other geometrically exact beam theory codes is that is written completely in the Julia programming language.  This presents several advantages for the `GXBeam` package. First, since Julia is a higher-level language, the code is generally easier to develop, maintain, and extend than lower-level languages.  This is especially helpful from a research perspective if one wishes to include `GXBeam` as a component of a multidisciplinary design optimization framework or fluid structure interaction solver.  Second, by leveraging Julia's type system, Julia-specific automatic differentiation packages (such as ForwardDiff [@Revels2016]) may be used to obtain exact derivatives for sensitivity analyses or gradient-based optimization.  Third, by maintaining type stability and minimizing allocations, this package is able to perform analyses with minimal overhead compared to lower-level languages such as C or Fortran.  Finally, the code is able to access and use several well-developed Julia-specific packages to enhance its capabilities such as NLsolve [@Mogensen2020] for solving nonlinear sets of equations, WriteVTK [@Polanco2021] for writing visualization files, and DifferentialEquations [@Rackauckas2017] for solving differential equations. 
 
