### 2022-01-21 e6c3f1c
final paper updates
