+++ paper/paper.md
@@ -33,6 +33,6 @@ One of the key advantages of `GXBeam` relative to other geometrically exact beam
 
 Even if one were to disregard the advantages associated with the use of the Julia language, `GXBeam` is still one of the most feature-rich open-source geometrically exact beam theory programs available.  Rather than restricting analyses to a single beam, `GXBeam` is able to model complex systems of interconnected nonlinear composite beams.  `GXBeam` also allows for a wide variety of analyses to be performed including linear or nonlinear static, steady state, eigenvalue, and time marching analyses.  Loads in `GXBeam` may be applied to nodes or elements and expressed as arbitrary functions of time.  Native support for gravitational loads and reference frame linear/angular velocities and accelerations are also supported.  Additionally, `GXBeam` allows point masses or rigid bodies with potentially time-varying inertial properties to be placed at arbitrary locations throughout each beam assembly.
 
-`GXBeam` may be used as either a standalone tool, or as a component of a larger analysis framework.  Its results are designed to be smooth and continuous, so that the package may be used as part of a gradient-based design optimization framework.  The package is also designed to be modular, so that it can be used as part of a fluid-structure interaction framework.  It has also been verified and validated using analytical, computational, and experimental results so that users may be reasonably confident that the results predicted by GXBeam are correct (when used within the theoretical limitations of geometrically exact beam theory).  These verifications and validations are included as part of the package's documentation so the accuracy of the package can be verified by anyone wishing to use it.  These features make GXBeam an invaluable tool for modeling highly flexible slender structures.
+`GXBeam` may be used as either a standalone tool, or as a component of a larger analysis framework.  Its results are designed to be smooth and continuous, so that the package may be used as part of a gradient-based design optimization framework.  The package is also designed to be modular, so that it can be used as part of a fluid-structure interaction framework.  It has also been verified and validated using analytical, computational, and experimental results so that users may be reasonably confident that the results predicted by GXBeam are correct (when used within the theoretical limitations of geometrically exact beam theory).  These verifications and validations are included as part of the package's documentation so the accuracy of the package can be verified by anyone wishing to use it.  These features make GXBeam an extremely useful tool for modeling the behavior of highly flexible slender structures.
 
 # References
