+++ paper/paper.md
@@ -33,6 +33,8 @@ One of the key advantages of `GXBeam` relative to other geometrically exact beam
 
 Even if one were to disregard the advantages associated with the use of the Julia language, `GXBeam` is still one of the most feature-rich open-source geometrically exact beam theory programs available.  Rather than restricting analyses to a single beam, `GXBeam` is able to model complex systems of interconnected nonlinear composite beams.  `GXBeam` also allows for a wide variety of analyses to be performed including linear or nonlinear static, steady state, eigenvalue, and time marching analyses.  Loads in `GXBeam` may be applied to nodes or elements and expressed as arbitrary functions of time.  Native support for gravitational loads and reference frame linear/angular velocities and accelerations are also supported.  Additionally, `GXBeam` allows point masses or rigid bodies with potentially time-varying inertial properties to be placed at arbitrary locations throughout each beam assembly.
 
-`GXBeam` may be used as either a standalone tool, or as a component of a larger analysis framework.  Its results are designed to be smooth and continuous, so that the package may be used as part of a gradient-based design optimization framework.  The package is also designed to be modular, so that it can be used as part of a fluid-structure interaction framework.  It has also been verified and validated using analytical, computational, and experimental results so that users may be reasonably confident that the results predicted by GXBeam are correct (when used within the theoretical limitations of geometrically exact beam theory).  These verifications and validations are included as part of the package's documentation so the accuracy of the package can be verified by anyone wishing to use it.  These features make GXBeam an extremely useful tool for modeling the behavior of highly flexible slender structures.
+While `GXBeam` may be applied to a large number of problems, it has some limitations which should be kept in mind.  First, by using the simplest possible shape functions (a combination of constant and linear shape functions), this package avoids using numerical quadrature except when integrating applied distributed loads. As a result, element properties are approximated as constant throughout each beam element and a relatively large number of beam elements may be necessary to achieve grid-independent results.  This package also does not currently model cross section warping and therefore should not be used to model open cross sections (such as I, C, or L-beams). The one exception to this rule is if a beam's width is much greater than its height, in which case the beam may be considered to be strip-like (like a helicopter blade).  `GXBeam` also relies on the results of linear cross-sectional analyses and therefore does not model the nonlinear component of the Trapeze effect, which is the tendency of a beam to untwist when subjected to axial tension. This nonlinear effect is typically most important when modeling rotating structures such as helicopter blades due to the presence of large centrifugal forces. It is also more important when modeling strip-like beams than for modeling closed cross-section beams due to their low torsional rigidity.
+
+Despite its current limitations, `GXBeam` is useful in a variety of situations both as a standalone tool and as a component of a larger analysis framework.  Its results are designed to be smooth and continuous, so that the package may be used as part of a gradient-based design optimization framework.  The package is also designed to be modular, so that it can be used as part of a fluid-structure interaction framework.  It has also been verified and validated using analytical, computational, and experimental results so that users may be reasonably confident that the results predicted by GXBeam are correct (when used within the theoretical limitations of geometrically exact beam theory).  These verifications and validations are included as part of the package's documentation so the accuracy of the package can be verified by anyone wishing to use it.  In summary, the feature-rich and validated capabilities of GXBeam make it a very useful tool in a variety of use cases for modeling the behavior of highly flexible slender structures.
 
 # References
