+++ paper/paper.bib
@@ -1,13 +1,13 @@
 @article{Bezanson2017,
  author = {Bezanson, Jeff and Edelman, Alan and Karpinski, Stefan and Shah, Viral B.},
- title = {Julia: {A} Fresh Approach to Numerical Computing},
+ title = {Julia: A Fresh Approach to Numerical Computing},
  number = {1},
  pages = {65--98},
  doi = {10.1137/141000671},
  url = {https://doi.org/10.1137/141000671},
  volume = {59},
  journal = {SIAM Rev.},
- publisher = {Society for Industrial \& Applied Mathematics (SIAM)},
+ publisher = {Society for Industrial \& Applied Mathematics ({SIAM})},
  year = {2017},
  source = {Crossref},
  issn = {0036-1445, 1095-7200},
@@ -15,344 +15,311 @@
 }
 
 @article{Revels2016,
- author = {{Revels}, J. and {Lubin}, M. and {Papamarkou}, T.},
- title = {Forward-Mode Automatic Differentiation in {{J}ulia}},
+ author = {Revels, J. and Lubin, M. and Papamarkou, T.},
+ title = {Forward-Mode Automatic Differentiation in {J}ulia},
  url = {https://arxiv.org/abs/1607.07892},
  journal = {arXiv:1607.07892 [cs.MS]},
  year = {2016},
 }
 
 @article{Hodges1990,
- author = {Hodges, Dewey H.},
  title = {A mixed variational formulation based on exact intrinsic equations for dynamics of moving beams},
- journal = {Int. J. Solids Struct.},
+ journal = {International Journal of Solids and Structures},
  volume = {26},
  number = {11},
- pages = {1253--1273},
+ pages = {1253-1273},
  year = {1990},
- publisher = {Elsevier BV},
- doi = {10.1016/0020-7683(90)90060-9},
- source = {Crossref},
- url = {https://doi.org/10.1016/0020-7683(90)90060-9},
  issn = {0020-7683},
+ doi = {10.1016/0020-7683(90)90060-9},
+ url = {https://doi.org/10.1016/0020-7683(90)90060-9},
+ author = {Dewey H. Hodges},
 }
 
-@article{Simo1986,
+@article{10.1115/1.3171871,
  author = {Simo, J. C. and Vu-Quoc, L.},
- title = {On the Dynamics of Flexible Beams Under Large Overall {Motions—The} Plane Case: {Part} {II}},
- year = {1986},
- doi = {10.1115/1.3171871},
- number = {4},
- source = {Crossref},
- url = {https://doi.org/10.1115/1.3171871},
+ title = "{On the Dynamics of Flexible Beams Under Large Overall Motions—The Plane Case: Part II}",
+ journal = {Journal of Applied Mechanics},
  volume = {53},
- journal = {J. Appl. Mech.},
- publisher = {ASME International},
- issn = {0021-8936, 1528-9036},
- pages = {855--863},
- month = dec,
+ number = {4},
+ pages = {855-863},
+ year = {1986},
+ month = {12},
+ issn = {0021-8936},
+ doi = {10.1115/1.3171871},
+ url = {https://doi.org/10.1115/1.3171871},
 }
 
 @article{Simo1988,
- author = {Simo, J.C. and Vu-Quoc, L.},
+ author = {J.C. Simo and L. Vu-Quoc},
  title = {On the dynamics in space of rods undergoing large motions — A geometrically exact approach},
- journal = {Comput. Method. Appl. M.},
+ journal = {Computer Methods in Applied Mechanics and Engineering},
  volume = {66},
  number = {2},
- pages = {125--161},
+ pages = {125-161},
  year = {1988},
[…]
