+++ paper/paper.bib
@@ -63,7 +63,7 @@
 }
 
 @article{Simo1985,
-author = {J. C. Simo}
+author = {J. C. Simo},
 title = {A finite strain beam formulation. The three-dimensional dynamic problem. Part {I}},
 journal = {Computer Methods in Applied Mechanics and Engineering},
 volume = {49},
