+++ paper/paper.bib
@@ -105,7 +105,7 @@ url = {https://doi.org/10.1016/0045-7825(85)90050-7},
   month        = dec,
   year         = 2020,
   doi          = {10.5281/zenodo.2682214},
-  url          = {https://doi.org/10.5281/zenodo.2682214}
+  url          = {https://doi.org/10.5281/zenodo.2682214},
   publisher    = {Zenodo},
 }
 
