+++ paper/paper.bib
@@ -294,7 +294,7 @@ month = jun,
  doi = {10.1002/nme.1620260710},
  url = {https://doi.org/10.1002/nme.1620260710},
  abstract = {Abstract An incremental Total Lagrangian Formulation for curved beam elements that includes the effect of large rotation increments is developed. A complete and symmetric tangent stiffness matrix is obtained and the numerical results show, in general, an improvement over the standard formulation where the assumption of infinitesimal rotation increments is made in the derivation of the tangent stiffness matrix.},
- year = {1988}
+ year = {1988},
  month = jul,
 }
 
