+++ paper/paper.bib
@@ -305,7 +305,7 @@ journal = {International Journal for Numerical Methods in Engineering},
 volume = {54},
 number = {12},
 pages = {1775--1788},
-year = {2002}
+year = {2002},
 doi = {10.1002/nme.487},
 url = {https://doi.org/10.1002/nme.487},
 }
