+++ paper/paper.bib
@@ -35,7 +35,7 @@
  author = {Dewey H. Hodges},
 }
 
-@article{10.1115/1.3171871,
+@article{Simo1986,
  author = {Simo, J. C. and Vu-Quoc, L.},
  title = "{On the Dynamics of Flexible Beams Under Large Overall Motions—The Plane Case: Part II}",
  journal = {Journal of Applied Mechanics},
