+++ paper/paper.bib
@@ -198,7 +198,7 @@ month = jun,
 
 @article{Ritto2002,
  author = {Ritto-Corr\^ea, M. and Camotim, D.},
- title = {On the differentiation of the Rodrigues formula and its significance for the vector-like parameterization of Reissner-Simo beam theory},
+ title = {On the differentiation of the {R}odrigues formula and its significance for the vector-like parameterization of {R}eissner-{S}imo beam theory},
  journal = {International Journal for Numerical Methods in Engineering},
  volume = {55},
  number = {9},
