+++ paper/paper.bib
@@ -37,7 +37,7 @@
 
 @article{Simo1986,
  author = {Simo, J. C. and Vu-Quoc, L.},
- title = "{On the Dynamics of Flexible Beams Under Large Overall Motions—The Plane Case: Part II}",
+ title = "{On the Dynamics of Flexible Beams Under Large Overall Motions--The Plane Case: Part II}",
  journal = {Journal of Applied Mechanics},
  volume = {53},
  number = {4},
@@ -51,7 +51,7 @@
 
 @article{Simo1988,
  author = {J.C. Simo and L. Vu-Quoc},
- title = {On the dynamics in space of rods undergoing large motions — A geometrically exact approach},
+ title = {On the dynamics in space of rods undergoing large motions -- A geometrically exact approach},
  journal = {Computer Methods in Applied Mechanics and Engineering},
  volume = {66},
  number = {2},
