### 2022-01-25 c32d4c6
add stiffness proportional damping
