+++ src/analyses.jl
@@ -106,8 +106,8 @@ function static_analysis!(system, assembly;
 
     # unpack pre-allocated storage and pointers
     x = system.x
-    F = system.r
-    J = system.K
+    resid = system.r
+    jacob = system.K
     force_scaling = system.force_scaling
     irow_point = system.irow_point
     irow_elem = system.irow_elem
@@ -129,10 +129,10 @@ function static_analysis!(system, assembly;
         gvec = typeof(gravity) <: AbstractVector ? SVector{3}(gravity) : SVector{3}(gravity(tvec[it]))
 
         # solve the system of equations
-        f! = (F, x) -> static_system_residual!(F, x, assembly, pcond, dload, pmass, gvec, force_scaling,
+        f! = (resid, x) -> static_system_residual!(resid, x, assembly, pcond, dload, pmass, gvec, force_scaling,
             irow_point, irow_elem1, irow_elem2, icol_point, icol_elem)
 
-        j! = (J, x) -> static_system_jacobian!(J, x, assembly, pcond, dload, pmass, gvec, force_scaling,
+        j! = (jacob, x) -> static_system_jacobian!(jacob, x, assembly, pcond, dload, pmass, gvec, force_scaling,
             irow_point, irow_elem1, irow_elem2, icol_point, icol_elem)
 
         if linear
@@ -144,12 +144,12 @@ function static_analysis!(system, assembly;
                     x .= linearization_state
                 end
             end
-            f!(F, x)
-            j!(J, x)
-            x .-= safe_lu(J) \ F
+            f!(resid, x)
+            j!(jacob, x)
+            x .-= safe_lu(jacob) \ resid
         else
             # nonlinear analysis
-            df = NLsolve.OnceDifferentiable(f!, j!, x, F, J)
+            df = NLsolve.OnceDifferentiable(f!, j!, x, resid, jacob)
 
             result = NLsolve.nlsolve(df, x,
                 linsolve=(x, A, b) -> ldiv!(x, safe_lu(A), b),
@@ -160,7 +160,7 @@ function static_analysis!(system, assembly;
 
             # update the solution
             x .= result.zero
-            J .= df.DF
+            jacob .= df.DF
 
             # update convergence flag
             converged = result.f_converged
@@ -304,8 +304,8 @@ function steady_state_analysis!(system, assembly;
 
     # unpack pointers to pre-allocated storage
     x = system.x
-    F = system.r
-    J = system.K
+    resid = system.r
+    jacob = system.K
     force_scaling = system.force_scaling
     irow_point = system.irow_point
     irow_elem = system.irow_elem
@@ -331,11 +331,11 @@ function steady_state_analysis!(system, assembly;
         a0 = typeof(linear_acceleration) <: AbstractVector ? SVector{3}(linear_acceleration) : SVector{3}(linear_acceleration(t))
         α0 = typeof(angular_acceleration) <: AbstractVector ? SVector{3}(angular_acceleration) : SVector{3}(angular_acceleration(t))
 
-        f! = (F, x) -> steady_state_system_residual!(F, x, assembly, pcond, dload, pmass, gvec, force_scaling, 
+        f! = (resid, x) -> steady_state_system_residual!(resid, x, assembly, pcond, dload, pmass, gvec, force_scaling, 
             irow_point, irow_elem, irow_elem1, irow_elem2, icol_point, 
             icol_elem, x0, v0, ω0, a0, α0)
 
[…]
