+++ src/system.jl
@@ -20,6 +20,8 @@ needed for time domain simulations.
  - `icol_elem`: Row/Column index of first state variable for each beam element
  - `udot`: Time derivative of state variable `u` for each beam element
  - `θdot`: Time derivative of state variable `θ` for each beam element
+ - `Fdot`: Time derivative of state variable `F` for each beam element
+ - `Mdot`: Time derivative of state variable `M` for each beam element
  - `Vdot`: Time derivative of state variable `V` for each beam element
  - `Ωdot`: Time derivative of state variable `Ω` for each beam element
  - `t`: Current system time
@@ -39,6 +41,8 @@ mutable struct System{TF, TV<:AbstractVector{TF}, TM<:AbstractMatrix{TF}}
     icol_elem::Vector{Int}
     udot::Vector{SVector{3,TF}}
     θdot::Vector{SVector{3,TF}}
+    Fdot::Vector{SVector{3,TF}}
+    Mdot::Vector{SVector{3,TF}}
     Vdot::Vector{SVector{3,TF}}
     Ωdot::Vector{SVector{3,TF}}
     t::TF
@@ -93,6 +97,8 @@ function System(TF, assembly, static;
     # initialize storage for time domain simulations
     udot = [@SVector zeros(TF, 3) for i = 1:nelem]
     θdot = [@SVector zeros(TF, 3) for i = 1:nelem]
+    Fdot = [@SVector zeros(TF, 3) for i = 1:nelem]
+    Mdot = [@SVector zeros(TF, 3) for i = 1:nelem]
     Vdot = [@SVector zeros(TF, 3) for i = 1:nelem]
     Ωdot = [@SVector zeros(TF, 3) for i = 1:nelem]
 
@@ -105,7 +111,7 @@ function System(TF, assembly, static;
 
     return System{TF, TV, TM}(static, x, r, K, M, force_scaling,
         irow_point, irow_elem, irow_elem1, irow_elem2, icol_point, icol_elem, udot, θdot,
-        Vdot, Ωdot, t)
+        Fdot, Mdot, Vdot, Ωdot, t)
 end
 
 function default_force_scaling(assembly)
@@ -969,7 +975,7 @@ end
 """
     initial_condition_system_residual!(resid, x, assembly, prescribed_conditions, distributed_loads, point_masses, gvec,
         force_scaling, irow_point, irow_elem, irow_elem1, irow_elem2, icol_point, icol_elem,
-        x0, v0, ω0, a0, α0, u, θ, udot, θdot)
+        x0, v0, ω0, a0, α0, u, θ, udot, θdot, Fdot, Mdot)
 
 Populate the system residual vector `resid` for an initial conditions analysis.
 
@@ -997,10 +1003,12 @@ Populate the system residual vector `resid` for an initial conditions analysis.
  - `θ`: initial angular deflections for each beam element
  - `udot`: initial linear deflection rates for each beam element
  - `θdot`: initial angular deflection rates for each beam element
+ - `Fdot`: initial elastic force rate for each beam element
+ - `Mdot`: initial elastic moment rate for each beam element
 """
 function initial_condition_system_residual!(resid, x, assembly, prescribed_conditions, distributed_loads, point_masses, gvec,
     force_scaling, irow_point, irow_elem, irow_elem1, irow_elem2, icol_point, icol_elem,
-    x0, v0, ω0, a0, α0, u, θ, udot, θdot)
+    x0, v0, ω0, a0, α0, u, θ, udot, θdot, Fdot, Mdot)
 
     npoint = length(assembly.points)
     nelem = length(assembly.elements)
[…]

[…]
