### 2022-01-25 290446f
put together damping example in tests
