### 2022-02-04 byuflowlab/GXBeam.jl#68 @Kevin-Mattheus-Moerman
Error testing GXBeam
I ran `] test GXBeam` and got the following error message:

```julia
┌ Warning: dt <= dtmin. Aborting. There is either an error in your model specification or the true solution is unstable.
└ @ SciMLBase ~/.julia/packages/SciMLBase/L7Nun/src/integrator_interface.jl:345
DifferentialEquations: Test Failed at /home/kevin/.julia/packages/GXBeam/S8Yc6/test/runtests.jl:1200
  Expression: sol.t[end] == 2.0
   Evaluated: 0.4756178300699102 == 2.0
Stacktrace:
 [1] macro expansion
   @ ~/.julia/packages/GXBeam/S8Yc6/test/runtests.jl:1200 [inlined]
 [2] macro expansion
   @ /buildworker/worker/package_linux64/build/usr/share/julia/stdlib/v1.6/Test/src/Test.jl:1151 [inlined]
 [3] top-level scope
   @ ~/.julia/packages/GXBeam/S8Yc6/test/runtests.jl:1141
DifferentialEquations: Error During Test at /home/kevin/.julia/packages/GXBeam/S8Yc6/test/runtests.jl:1139
  Got exception outside of a @test
  type Nothing has no field α
  Stacktrace:
    [1] setproperty!(x::Nothing, f::Symbol, v::Float64)
      @ Base ./Base.jl:34
    [2] calc_W!
      @ ~/.julia/packages/OrdinaryDiffEq/Op0Oq/src/derivative_utils.jl:549 [inlined]
    [3] update_W!
      @ ~/.julia/packages/OrdinaryDiffEq/Op0Oq/src/derivative_utils.jl:653 [inlined]
    [4] nlsolve!(nlsolver::OrdinaryDiffEq.NLSolver{NLNewton{Rational{Int64}, Rational{Int64}, Rational{Int64}}, true, Vector{Float64}, Rational{Int64}, Nothing, Float64, OrdinaryDiffEq.NLNewtonCache{Vector{Float64}, Float64, Float64, Vector{Float64}, Matrix{Float64}, Matrix{Float64}, Nothing, Nothing, LinearSolve.LinearCache{Matrix{Float64}, Vector{Float64}, Vector{Float64}, SciMLBase.NullParameters, Nothing, LU{Float64, Matrix{Float64}}, LinearSolve.InvPreconditioner{Diagonal{Float64, Vector{Float64}}}, Diagonal{Float64, Vector{Float64}}, Float64}}}, integrator::OrdinaryDiffEq.ODEIntegrator{DABDF2{12, true, Nothing, NLNewton{Rational{Int64}, Rational{Int64}, Rational{Int64}}, typeof(OrdinaryDiffEq.DEFAULT_PRECS), Val{:forward}, true, nothing}, true, Vector{Float64}, Vector{Float64}, Float64, Tuple{var"#103#105"{Int64}, Dict{Int64, DistributedLoads{Float64}}, Dict{Int64, Vector{PointMass{Float64}}}, StaticArrays.SVector{3, Float64}, StaticArrays.SVector{3, Float64}, StaticArrays.SVector{3, Float64}, StaticArrays.SVector{3, Float64}, StaticArrays.SVector{3, Float64}, StaticArrays.SVector{3, Float64}}, Float64, Float64, Float64, Float64, Vector{Vector{Float64}}, DAESolution{Float64, 2, Vector{Vector{Float64}}, Nothing, Nothing, Nothing, Vector{Float64}, DAEProblem{Vector{Float64}, Vector{Float64}, Tuple{Float64, Float64}, true, Tuple{var"#103#105"{Int64}, Dict{Int64, DistributedLoads{Float64}}, Dict{Int64, Vector{PointMass{Float64}}}, StaticArrays.SVector{3, Float64}, StaticArrays.SVector{3, Float64}, StaticArrays.SVector{3, Float64}, StaticArrays.SVector{3, Float64}, StaticArrays.SVector{3, Float64}, StaticArrays.SVector{3, Float64}}, DAEFunction{true, GXBeam.var"#299#301"{Assembly{Float64, Vector{Vector{Float64}}, UnitRange{Int64}, Vector{GXBeam.Element{Float64}}}, Float64, Vector{Int64}, Vector{Int64}, Vector{Int64}, Vector{Int64}, Vector{Int64}, Vector{Int64}}, Nothing, Nothing, GXBeam.var"#300#302"{Assembly{Float64, Vector{Vector{Float64}}, UnitRange{Int64}, Vector{GXBeam.Element{Float64}}}, Float64, Vector{Int64}, Vector{Int64}, Vector{Int64}, Vector{Int64}, Vector{Int64}, Vector{Int64}}, Nothing, Nothing, Nothing, SparseArrays.SparseMatrixCSC{Bool, Int64}, Nothing, Nothing, Nothing, Nothing, Nothing}, Base.Iterators.Pairs{Union{}, Union{}, Tuple{}, NamedTuple{(), Tuple{}}}, Vector{Bool}}, DABDF2{12, true, Nothing, NLNewton{Rational{Int64}, Rational{Int64}, Rational{Int64}}, typeof(OrdinaryDiffEq.DEFAULT_PRECS), Val{:forward}, true, nothing}, OrdinaryDiffEq.InterpolationData{DAEFunction{true, GXBeam.var"#299#301"{Assembly{Float64, Vector{Vector{Float64}}, UnitRange{Int64}, Vector{GXBeam.Element{Float64}}}, Float64, Vector{Int64}, Vector{Int64}, Vector{Int64}, Vector{Int64}, Vector{Int64}, Vector{Int64}}, Nothing, Nothing, GXBeam.var"#300#302"{Assembly{Float64, Vector{Vector{Float64}}, UnitRange{Int64}, Vector{GXBeam.Element{Float64}}}, Float64, Vector{Int64}, Vector{Int64}, Vector{Int64}, Vector{Int64}, Vector{Int64}, Vector{Int64}}, Nothing, Nothing, Nothing, SparseArrays.SparseMatrixCSC{Bool, Int64}, Nothing, Nothing, Nothing, Nothing, Nothing}, Vector{Vector{Float64}}, Vector{Float64}, Vector{Vector{Vector{Float64}}}, OrdinaryDiffEq.DABDF2Cache{Vector{Float64}, Vector{Float64}, Vector{Float64}, OrdinaryDiffEq.NLSolver{NLNewton{Rational{Int64}, Rational{Int64}, Rational{Int64}}, true, Vector{Float64}, Rational{Int64}, Nothing, Float64, OrdinaryDiffEq.NLNewtonCache{Vector{Float64}, Float64, Float64, Vector{Float64}, Matrix{Float64}, Matrix{Float64}, Nothing, Nothing, LinearSolve.LinearCache{Matrix{Float64}, Vector{Float64}, Vector{Float64}, SciMLBase.NullParameters, Nothing, LU{Float64, Matrix{Float64}}, LinearSolve.InvPreconditioner{Diagonal{Float64, Vector{Float64}}}, Diagonal{Float64, Vector{Float64}}, Float64}}}, Float64}}, DiffEqBase.DEStats}, DAEFunction{true, GXBeam.var"#299#301"{Assembly{Float64, Vector{Vector{Float64}}, UnitRange{Int64}, Vector{GXBeam.Element{Float64}}}, Float64, Vector{Int64}, Vector{Int64}, Vector{Int64}, Vector{Int64}, Vector{Int64}, Vector{Int64}}, Nothing, Nothing, GXBeam.var"#300#302"{Assembly{Float64, Vector{Vector{Float64}}, UnitRange{Int64}, Vector{GXBeam.Element{Float64}}}, Float64, Vector{Int64}, Vector{Int64}, Vector{Int64}, Vector{Int64}, Vector{Int64}, Vector{Int64}}, Nothing, Nothing, Nothing, SparseArrays.SparseMatrixCSC{Bool, Int64}, Nothing, Nothing, Nothing, Nothing, Nothing}, OrdinaryDiffEq.DABDF2Cache{Vector{Float64}, Vector{Float64}, Vector{Float64}, OrdinaryDiffEq.NLSolver{NLNewton{Rational{Int64}, Rational{Int64}, Rational{Int64}}, true, Vector{Float64}, Rational{Int64}, Nothing, Float64, OrdinaryDiffEq.NLNewtonCache{Vector{Float64}, Float64, Float64, Vector{Float64}, Matrix{Float64}, Matrix{Float64}, Nothing, Nothing, LinearSolve.LinearCache{Matrix{Float64}, Vector{Float64}, Vector{Float64}, SciMLBase.NullParameters, Nothing, LU{Float64, Matrix{Float64}}, LinearSolve.InvPreconditioner{Diagonal{Float64, Vector{Float64}}}, Diagonal{Float64, Vector{Float64}}, Float64}}}, Float64}, OrdinaryDiffEq.DEOptions{Float64, Float64, Float64, Float64, PIController{Rational{Int64}}, typeof(DiffEqBase.ODE_DEFAULT_NORM), typeof(opnorm), Nothing, CallbackSet{Tuple{}, Tuple{}}, typeof(DiffEqBase.ODE_DEFAULT_ISOUTOFDOMAIN), typeof(DiffEqBase.ODE_DEFAULT_PROG_MESSAGE), typeof(DiffEqBase.ODE_DEFAULT_UNSTABLE_CHECK), DataStructures.BinaryHeap{Float64, DataStructures.FasterForward}, DataStructures.BinaryHeap{Float64, DataStructures.FasterForward}, Nothing, Nothing, Int64, Tuple{}, Tuple{}, Tuple{}}, Vector{Float64}, Float64, Nothing, OrdinaryDiffEq.DefaultInit}, cache::OrdinaryDiffEq.DImplicitEulerCache{Vector{Float64}, Vector{Float64}, Vector{Float64}, OrdinaryDiffEq.NLSolver{NLNewton{Rational{Int64}, Rational{Int64}, Rational{Int64}}, true, Vector{Float64}, Rational{Int64}, Nothing, Float64, OrdinaryDiffEq.NLNewtonCache{Vector{Float64}, Float64, Float64, Vector{Float64}, Matrix{Float64}, Matrix{Float64}, Nothing, Nothing, LinearSolve.LinearCache{Matrix{Float64}, Vector{Float64}, Vector{Float64}, SciMLBase.NullParameters, Nothing, LU{Float64, Matrix{Float64}}, LinearSolve.InvPreconditioner{Diagonal{Float64, Vector{Float64}}}, Diagonal{Float64, Vector{Float64}}, Float64}}}}, repeat_step::Bool)
      @ OrdinaryDiffEq ~/.julia/packages/OrdinaryDiffEq/Op0Oq/src/nlsolve/nlsolve.jl:21
    [5] perform_step!(integrator::OrdinaryDiffEq.ODEIntegrator{DABDF2{12, true, Nothing, NLNewton{Rational{Int64}, Rational{Int64}, Rational{Int64}}, typeof(OrdinaryDiffEq.DEFAULT_PRECS), Val{:forward}, true, nothing}, true, Vector{Float64}, Vector{Float64}, Float64, Tuple{var"#103#105"{Int64}, Dict{Int64, DistributedLoads{Float64}}, Dict{Int64, Vector{PointMass{Float64}}}, StaticArrays.SVector{3, Float64}, StaticArrays.SVector{3, Float64}, StaticArrays.SVector{3, Float64}, StaticArrays.SVector{3, Float64}, StaticArrays.SVector{3, Float64}, StaticArrays.SVector{3, Float64}}, Float64, Float64, Float64, Float64, Vector{Vector{Float64}}, DAESolution{Float64, 2, Vector{Vector{Float64}}, Nothing, Nothing, Nothing, Vector{Float64}, DAEProblem{Vector{Float64}, Vector{Float64}, Tuple{Float64, Float64}, true, Tuple{var"#103#105"{Int64}, Dict{Int64, DistributedLoads{Float64}}, Dict{Int64, Vector{PointMass{Float64}}}, StaticArrays.SVector{3, Float64}, StaticArrays.SVector{3, Float64}, StaticArrays.SVector{3, Float64}, StaticArrays.SVector{3, Float64}, StaticArrays.SVector{3, Float64}, StaticArrays.SVector{3, Float64}}, DAEFunction{true, GXBeam.var"#299#301"{Assembly{Float64, Vector{Vector{Float64}}, UnitRange{Int64}, Vector{GXBeam.Element{Float64}}}, Float64, Vector{Int64}, Vector{Int64}, Vector{Int64}, Vector{Int64}, Vector{Int64}, Vector{Int64}}, Nothing, Nothing, GXBeam.var"#300#302"{Assembly{Float64, Vector{Vector{Float64}}, UnitRange{Int64}, Vector{GXBeam.Element{Float64}}}, Float64, Vector{Int64}, Vector{Int64}, Vector{Int64}, Vector{Int64}, Vector{Int64}, Vector{Int64}}, Nothing, Nothing, Nothing, SparseArrays.SparseMatrixCSC{Bool, Int64}, Nothing, Nothing, Nothing, Nothing, Nothing}, Base.Iterators.Pairs{Union{}, Union{}, Tuple{}, NamedTuple{(), Tuple{}}}, Vector{Bool}}, DABDF2{12, true, Nothing, NLNewton{Rational{Int64}, Rational{Int64}, Rational{Int64}}, typeof(OrdinaryDiffEq.DEFAULT_PRECS), Val{:forward}, true, nothing}, OrdinaryDiffEq.InterpolationData{DAEFunction{true, GXBeam.var"#299#301"{Assembly{Float64, Vector{Vector{Float64}}, UnitRange{Int64}, Vector{GXBeam.Element{Float64}}}, Float64, Vector{Int64}, Vector{Int64}, Vector{Int64}, Vector{Int64}, Vector{Int64}, Vector{Int64}}, Nothing, Nothing, GXBeam.var"#300#302"{Assembly{Float64, Vector{Vector{Float64}}, UnitRange{Int64}, Vector{GXBeam.Element{Float64}}}, Float64, Vector{Int64}, Vector{Int64}, Vector{Int64}, Vector{Int64}, Vector{Int64}, Vector{Int64}}, Nothing, Nothing, Nothing, SparseArrays.SparseMatrixCSC{Bool, Int64}, Nothing, Nothing, Nothing, Nothing, Nothing}, Vector{Vector{Float64}}, Vector{Float64}, Vector{Vector{Vector{Float64}}}, OrdinaryDiffEq.DABDF2Cache{Vector{Float64}, Vector{Float64}, Vector{Float64}, OrdinaryDiffEq.NLSolver{NLNewton{Rational{Int64}, Rational{Int64}, Rational{Int64}}, true, Vector{Float64}, Rational{Int64}, Nothing, Float64, OrdinaryDiffEq.NLNewtonCache{Vector{Float64}, Float64, Float64, Vector{Float64}, Matrix{Float64}, Matrix{Float64}, Nothing, Nothing, LinearSolve.LinearCache{Matrix{Float64}, Vector{Float64}, Vector{Float64}, SciMLBase.NullParameters, Nothing, LU{Float64, Matrix{Float64}}, LinearSolve.InvPreconditioner{Diagonal{Float64, Vector{Float64}}}, Diagonal{Float64, Vector{Float64}}, Float64}}}, Float64}}, DiffEqBase.DEStats}, DAEFunction{true, GXBeam.var"#299#301"{Assembly{Float64, Vector{Vector{Float64}}, UnitRange{Int64}, Vector{GXBeam.Element{Float64}}}, Float64, Vector{Int64}, Vector{Int64}, Vector{Int64}, Vector{Int64}, Vector{Int64}, Vector{Int64}}, Nothing, Nothing, GXBeam.var"#300#302"{Assembly{Float64, Vector{Vector{Float64}}, UnitRange{Int64}, Vector{GXBeam.Element{Float64}}}, Float64, Vector{Int64}, Vector{Int64}, Vector{Int64}, Vector{Int64}, Vector{Int64}, Vector{Int64}}, Nothing, Nothing, Nothing, SparseArrays.SparseMatrixCSC{Bool, Int64}, Nothing, Nothing, Nothing, Nothing, Nothing}, OrdinaryDiffEq.DABDF2Cache{Vector{Float64}, Vector{Float64}, Vector{Float64}, OrdinaryDiffEq.NLSolver{NLNewton{Rational{Int64}, Rational{Int64}, Rational{Int64}}, true, Vector{Float64}, Rational{Int64}, Nothing, Float64, OrdinaryDiffEq.NLNewtonCache{Vector{Float64}, Float64, Float64, Vector{Float64}, Matrix{Float64}, Matrix{Float64}, Nothing, Nothing, LinearSolve.LinearCache{Matrix{Float64}, Vector{Float64}, Vector{Float64}, SciMLBase.NullParameters, Nothing, LU{Float64, Matrix{Float64}}, LinearSolve.InvPreconditioner{Diagonal{Float64, Vector{Float64}}}, Diagonal{Float64, Vector{Float64}}, Float64}}}, Float64}, OrdinaryDiffEq.DEOptions{Float64, Float64, Float64, Float64, PIController{Rational{Int64}}, typeof(DiffEqBase.ODE_DEFAULT_NORM), typeof(opnorm), Nothing, CallbackSet{Tuple{}, Tuple{}}, typeof(DiffEqBase.ODE_DEFAULT_ISOUTOFDOMAIN), typeof(DiffEqBase.ODE_DEFAULT_PROG_MESSAGE), typeof(DiffEqBase.ODE_DEFAULT_UNSTABLE_CHECK), DataStructures.BinaryHeap{Float64, DataStructures.FasterForward}, DataStructures.BinaryHeap{Float64, DataStructures.FasterForward}, Nothing, Nothing, Int64, Tuple{}, Tuple{}, Tuple{}}, Vector{Float64}, Float64, Nothing, OrdinaryDiffEq.DefaultInit}, cache::OrdinaryDiffEq.DImplicitEulerCache{Vector{Float64}, Vector{Float64}, Vector{Float64}, OrdinaryDiffEq.NLSolver{NLNewton{Rational{Int64}, Rational{Int64}, Rational{Int64}}, true, Vector{Float64}, Rational{Int64}, Nothing, Float64, OrdinaryDiffEq.NLNewtonCache{Vector{Float64}, Float64, Float64, Vector{Float64}, Matrix{Float64}, Matrix{Float64}, Nothing, Nothing, LinearSolve.LinearCache{Matrix{Float64}, Vector{Float64}, Vector{Float64}, SciMLBase.NullParameters, Nothing, LU{Float64, Matrix{Float64}}, LinearSolve.InvPreconditioner{Diagonal{Float64, Vector{Float64}}}, Diagonal{Float64, Vector{Float64}}, Float64}}}}, repeat_step::Bool)
      @ OrdinaryDiffEq ~/.julia/packages/OrdinaryDiffEq/Op0Oq/src/perform_step/dae_perform_step.jl:67
    [6] perform_step!(integrator::OrdinaryDiffEq.ODEIntegrator{DABDF2{12, true, Nothing, NLNewton{Rational{Int64}, Rational{Int64}, Rational{Int64}}, typeof(OrdinaryDiffEq.DEFAULT_PRECS), Val{:forward}, true, nothing}, true, Vector{Float64}, Vector{Float64}, Float64, Tuple{var"#103#105"{Int64}, Dict{Int64, DistributedLoads{Float64}}, Dict{Int64, Vector{PointMass{Float64}}}, StaticArrays.SVector{3, Float64}, StaticArrays.SVector{3, Float64}, StaticArrays.SVector{3, Float64}, StaticArrays.SVector{3, Float64}, StaticArrays.SVector{3, Float64}, StaticArrays.SVector{3, Float64}}, Float64, Float64, Float64, Float64, Vector{Vector{Float64}}, DAESolution{Float64, 2, Vector{Vector{Float64}}, Nothing, Nothing, Nothing, Vector{Float64}, DAEProblem{Vector{Float64}, Vector{Float64}, Tuple{Float64, Float64}, true, Tuple{var"#103#105"{Int64}, Dict{Int64, DistributedLoads{Float64}}, Dict{Int64, Vector{PointMass{Float64}}}, StaticArrays.SVector{3, Float64}, StaticArrays.SVector{3, Float64}, StaticArrays.SVector{3, Float64}, StaticArrays.SVector{3, Float64}, StaticArrays.SVector{3, Float64}, StaticArrays.SVector{3, Float64}}, DAEFunction{true, GXBeam.var"#299#301"{Assembly{Float64, Vector{Vector{Float64}}, UnitRange{Int64}, Vector{GXBeam.Element{Float64}}}, Float64, Vector{Int64}, Vector{Int64}, Vector{Int64}, Vector{Int64}, Vector{Int64}, Vector{Int64}}, Nothing, Nothing, GXBeam.var"#300#302"{Assembly{Float64, Vector{Vector{Float64}}, UnitRange{Int64}, Vector{GXBeam.Element{Float64}}}, Float64, Vector{Int64}, Vector{Int64}, Vector{Int64}, Vector{Int64}, Vector{Int64}, Vector{Int64}}, Nothing, Nothing, Nothing, SparseArrays.SparseMatrixCSC{Bool, Int64}, Nothing, Nothing, Nothing, Nothing, Nothing}, Base.Iterators.Pairs{Union{}, Union{}, Tuple{}, NamedTuple{(), Tuple{}}}, Vector{Bool}}, DABDF2{12, true, Nothing, NLNewton{Rational{Int64}, Rational{Int64}, Rational{Int64}}, typeof(OrdinaryDiffEq.DEFAULT_PRECS), Val{:forward}, true, nothing}, OrdinaryDiffEq.InterpolationData{DAEFunction{true, GXBeam.var"#299#301"{Assembly{Float64, Vector{Vector{Float64}}, UnitRange{Int64}, Vector{GXBeam.Element{Float64}}}, Float64, Vector{Int64}, Vector{Int64}, Vector{Int64}, Vector{Int64}, Vector{Int64}, Vector{Int64}}, Nothing, Nothing, GXBeam.var"#300#302"{Assembly{Float64, Vector{Vector{Float64}}, UnitRange{Int64}, Vector{GXBeam.Element{Float64}}}, Float64, Vector{Int64}, Vector{Int64}, Vector{Int64}, Vector{Int64}, Vector{Int64}, Vector{Int64}}, Nothing, Nothing, Nothing, SparseArrays.SparseMatrixCSC{Bool, Int64}, Nothing, Nothing, Nothing, Nothing, Nothing}, Vector{Vector{Float64}}, Vector{Float64}, Vector{Vector{Vector{Float64}}}, OrdinaryDiffEq.DABDF2Cache{Vector{Float64}, Vector{Float64}, Vector{Float64}, OrdinaryDiffEq.NLSolver{NLNewton{Rational{Int64}, Rational{Int64}, Rational{Int64}}, true, Vector{Float64}, Rational{Int64}, Nothing, Float64, OrdinaryDiffEq.NLNewtonCache{Vector{Float64}, Float64, Float64, Vector{Float64}, Matrix{Float64}, Matrix{Float64}, Nothing, Nothing, LinearSolve.LinearCache{Matrix{Float64}, Vector{Float64}, Vector{Float64}, SciMLBase.NullParameters, Nothing, LU{Float64, Matrix{Float64}}, LinearSolve.InvPreconditioner{Diagonal{Float64, Vector{Float64}}}, Diagonal{Float64, Vector{Float64}}, Float64}}}, Float64}}, DiffEqBase.DEStats}, DAEFunction{true, GXBeam.var"#299#301"{Assembly{Float64, Vector{Vector{Float64}}, UnitRange{Int64}, Vector{GXBeam.Element{Float64}}}, Float64, Vector{Int64}, Vector{Int64}, Vector{Int64}, Vector{Int64}, Vector{Int64}, Vector{Int64}}, Nothing, Nothing, GXBeam.var"#300#302"{Assembly{Float64, Vector{Vector{Float64}}, UnitRange{Int64}, Vector{GXBeam.Element{Float64}}}, Float64, Vector{Int64}, Vector{Int64}, Vector{Int64}, Vector{Int64}, Vector{Int64}, Vector{Int64}}, Nothing, Nothing, Nothing, SparseArrays.SparseMatrixCSC{Bool, Int64}, Nothing, Nothing, Nothing, Nothing, Nothing}, OrdinaryDiffEq.DABDF2Cache{Vector{Float64}, Vector{Float64}, Vector{Float64}, OrdinaryDiffEq.NLSolver{NLNewton{Rational{Int64}, Rational{Int64}, Rational{Int64}}, true, Vector{Float64}, Rational{Int64}, Nothing, Float64, OrdinaryDiffEq.NLNewtonCache{Vector{Float64}, Float64, Float64, Vector{Float64}, Matrix{Float64}, Matrix{Float64}, Nothing, Nothing, LinearSolve.LinearCache{Matrix{Float64}, Vector{Float64}, Vector{Float64}, SciMLBase.NullParameters, Nothing, LU{Float64, Matrix{Float64}}, LinearSolve.InvPreconditioner{Diagonal{Float64, Vector{Float64}}}, Diagonal{Float64, Vector{Float64}}, Float64}}}, Float64}, OrdinaryDiffEq.DEOptions{Float64, Float64, Float64, Float64, PIController{Rational{Int64}}, typeof(DiffEqBase.ODE_DEFAULT_NORM), typeof(opnorm), Nothing, CallbackSet{Tuple{}, Tuple{}}, typeof(DiffEqBase.ODE_DEFAULT_ISOUTOFDOMAIN), typeof(DiffEqBase.ODE_DEFAULT_PROG_MESSAGE), typeof(DiffEqBase.ODE_DEFAULT_UNSTABLE_CHECK), DataStructures.BinaryHeap{Float64, DataStructures.FasterForward}, DataStructures.BinaryHeap{Float64, DataStructures.FasterForward}, Nothing, Nothing, Int64, Tuple{}, Tuple{}, Tuple{}}, Vector{Float64}, Float64, Nothing, OrdinaryDiffEq.DefaultInit}, cache::OrdinaryDiffEq.DABDF2Cache{Vector{Float64}, Vector{Float64}, Vector{Float64}, OrdinaryDiffEq.NLSolver{NLNewton{Rational{Int64}, Rational{Int64}, Rational{Int64}}, true, Vector{Float64}, Rational{Int64}, Nothing, Float64, OrdinaryDiffEq.NLNewtonCache{Vector{Float64}, Float64, Float64, Vector{Float64}, Matrix{Float64}, Matrix{Float64}, Nothing, Nothing, LinearSolve.LinearCache{Matrix{Float64}, Vector{Float64}, Vector{Float64}, SciMLBase.NullParameters, Nothing, LU{Float64, Matrix{Float64}}, LinearSolve.InvPreconditioner{Diagonal{Float64, Vector{Float64}}}, Diagonal{Float64, Vector{Float64}}, Float64}}}, Float64}, repeat_step::Bool)
      @ OrdinaryDiffEq ~/.julia/packages/OrdinaryDiffEq/Op0Oq/src/perform_step/dae_perform_step.jl:179
    [7] perform_step!
      @ ~/.julia/packages/OrdinaryDiffEq/Op0Oq/src/perform_step/dae_perform_step.jl:171 [inlined]
[…]
