### 2022-02-04 #3997 @Kevin-Mattheus-Moerman
@taylormcd I've started the review process. Here are some initial comments. I'll resume work early next week.

- [ ] On the [documentation site]() under `Limitations` it says `..More details about the convergence of this package may be found in the examples...`. The example link there is broken (it goes here which gives a 404: https://flow.byu.edu/GXBeam.jl/dev/@ref%20Nonlinear%20Analysis%20of%20a%20Cantilever%20Subjected%20to%20a%20Constant%20Moment)
- [ ] The examples in the documentation appear encoded in a mardown file (i.e. https://github.com/byuflowlab/GXBeam.jl/blob/master/docs/src/examples.md). It would be great however if these could be executed e.g. as jupyter notebooks. Could the authors consider facilitating this? It should largely be a copy-paste excercise to make these executable Jupyter notebooks, whereas currently, as a new user, I find myself having to copy paste a lot to reproduce the examples.
- [ ] I got an error running the tests, can you assist/investigate? https://github.com/byuflowlab/GXBeam.jl/issues/68
