### 2022-02-04 byuflowlab/GXBeam.jl#68 @taylormcd
This is odd considering the CI tests pass, though I also recently observed this behavior in the development branches documentation.  It's probably just a poorly posed DAE problem, at least when used with the DifferentialEquations solvers.  I'll look into this.
