### 2022-02-04 byuflowlab/GXBeam.jl#68 @taylormcd
Looks like the issue happens when using the latest version of OrdinaryDiffEq.  I'll probably just cap the version until I can get this resolved.
