### 2022-02-04 86cf8b2
fix diffeq documentation
