+++ .github/workflows/test.yaml
@@ -7,7 +7,7 @@ jobs:
     runs-on: ${{ matrix.os }}
     strategy:
       matrix:
-        julia-version: ['1.5']
+        julia-version: ['1.5', '1']
         julia-arch: [x64]
         os: [ubuntu-latest, windows-latest, macOS-latest]
 
