### 2022-02-07 #3997 @taylormcd
@Kevin-Mattheus-Moerman
 - The broken link is fixed now.
 - I used ![Documenter.jl](https://github.com/JuliaDocs/Documenter.jl) for the documentation and examples, which is pretty standard practice in Julia.  I would rather not have two versions of the same examples due to the headaches that causes for maintenance.  That being said, using ![Literate.jl](https://github.com/fredrikekre/Literate.jl) might be a good long term solution.
 - The error is now fixed.
