### 2022-03-03 cbe1cf2
calculate thetadot explicitly
