+++ src/element.jl
@@ -199,8 +199,9 @@ Calculate the element resultants for a steady state analysis.
     f_ψ1 += tmp
     f_ψ2 += tmp
 
+    Qinv = get_Qinv(θ)
     f_V = CtCab*V - v - cross(ω, u)
-    f_Ω = Ω - CtCab'*ω
+    f_Ω = Qinv*Cab*(Ω - CtCab'*ω)
 
     return f_u1, f_u2, f_ψ1, f_ψ2, f_F1, f_F2, f_M1, f_M2, f_V, f_Ω
 end
@@ -255,8 +256,7 @@ Calculate the element resultants for a dynamic analysis.
 
     f_V -= udot
 
-    Q = get_Q(θ)
-    f_Ω -= Cab'*Q*θdot
+    f_Ω -= θdot
 
     return f_u1, f_u2, f_ψ1, f_ψ2, f_F1, f_F2, f_M1, f_M2, f_V, f_Ω
 end
@@ -1123,7 +1123,7 @@ end
 
 """
     steady_state_element_jacobian_equations(ΔL, S11, S12, S21, S22, 
-        mass11, mass12, mass21, mass22, Cab, CtCab, θ, F, M, γ, κ, V, P, H, ω, f1_u, f2_u, 
+        mass11, mass12, mass21, mass22, Cab, CtCab, θ, F, M, γ, κ, V, Ω, P, H, ω, f1_u, f2_u, 
         m1_u, m2_u, f1_θ, f2_θ, m1_θ, m2_θ, C_θ1, C_θ2, C_θ3, Ct_θ1, Ct_θ2, Ct_θ3)
 
 Calculate the derivatives of the element resultants with respect to the state variables for 
@@ -1141,6 +1141,7 @@ a steady state analysis.
  - `γ`: engineering strain
  - `κ`: curvature
  - `V`: linear velocity
+ - `Ω`: angular velocity
  - `P`: linear momentum
  - `H`: angular momentum
  - `ω`: angular velocity caused by body frame velocity
@@ -1160,7 +1161,7 @@ a steady state analysis.
  - `Ct_θ3`: Derivative of `C'` w.r.t. `θ[3]` (transpose of `C_θ3`)
 """
 @inline function steady_state_element_jacobian_equations(ΔL, S11, S12, S21, S22, 
-    mass11, mass12, mass21, mass22, Cab, CtCab, θ, F, M, γ, κ, V, P, H, ω, f1_u, f2_u, 
+    mass11, mass12, mass21, mass22, Cab, CtCab, θ, F, M, γ, κ, V, Ω, P, H, ω, f1_u, f2_u, 
     m1_u, m2_u, f1_θ, f2_θ, m1_θ, m2_θ, C_θ1, C_θ2, C_θ3, Ct_θ1, Ct_θ2, Ct_θ3)
 
     f_u1_u, f_u2_u, f_u1_θ, f_u2_θ, f_u1_F, f_u2_F,
@@ -1220,11 +1221,15 @@ a steady state analysis.
 
     # --- f_Ω --- #
 
+    Qinv = get_Qinv(θ)
+    Qinv_θ1, Qinv_θ2, Qinv_θ3 = get_Qinv_θ(θ)
+
     # d_fH_dθ
-    f_Ω_θ = -Cab'*mul3(C_θ1, C_θ2, C_θ3, ω)
[…]
