+++ src/interfaces/differentialequations.jl
@@ -4,8 +4,52 @@
 Construct a `ODEProblem` for the system of nonlinear beams
 contained in `assembly` which may be used with the DifferentialEquations package.
 
-**Note that this function defines a non-constant mass matrix, which is not directly 
-supported by DifferentialEquations.jl.  This function is therefore experimental**
+Keyword Arguments:
+ - `prescribed_conditions = Dict{Int,PrescribedConditions{Float64}}()`:
+        A dictionary with keys corresponding to the points at
+        which prescribed conditions are applied and elements of type
+        [`PrescribedConditions`](@ref) which describe the prescribed conditions
+        at those points.  If time varying, this input may be provided as a
+        function of time.
+ - `distributed_loads = Dict{Int,DistributedLoads{Float64}}()`: A dictionary
+        with keys corresponding to the elements to which distributed loads are
+        applied and elements of type [`DistributedLoads`](@ref) which describe
+        the distributed loads on those elements.  If time varying, this input may
+        be provided as a function of time.
+ - `point_masses = Dict{Int,Vector{PointMass{Float64}}}()`: A dictionary with keys 
+        corresponding to the points at which point masses are attached and values 
+        containing vectors of objects of type [`PointMass`](@ref) which describe 
+        the point masses attached at those points.  If time varying, this input may
+        be provided as a function of time.
+ - `gravity`: Gravity vector. If time varying, this input may be provided as a 
+        function of time.
+ - `origin = zeros(3)`: Global frame origin vector. If time varying, this input
+        may be provided as a function of time.
+ - `linear_velocity = zeros(3)`: Global frame linear velocity vector. If time
+        varying, this vector may be provided as a function of time.
+ - `angular_velocity = zeros(3)`: Global frame angular velocity vector. If time
+        varying, this vector may be provided as a function of time.
+ - `linear_acceleration = zeros(3)`: Global frame linear acceleration vector. If time
+       varying, this vector may be provided as a function of time.
+ - `angular_acceleration = zeros(3)`: Global frame angular acceleration vector. If time
+       varying, this vector may be provided as a function of time.
+"""
+function SciMLBase.ODEProblem(system::System, assembly, tspan; kwargs...)
+
+    # use initial state from `system`
+    u0 = vcat(system.x, zeros(12*length(assembly.elements)))
+
+    # create ODEFunction
+    func = SciMLBase.ODEFunction(system, assembly; kwargs...)
+
+    return SciMLBase.ODEProblem{true}(func, u0, tspan)
+end
+
+"""
+    ODEFunction(system::GXBeam.System, assembly; kwargs...)
+
+Construct a `ODEFunction` for the system of nonlinear beams contained in `assembly` which 
+may be used with the DifferentialEquations package.
 
 Keyword Arguments:
  - `prescribed_conditions = Dict{Int,PrescribedConditions{Float64}}()`:
@@ -37,7 +81,7 @@ Keyword Arguments:
  - `angular_acceleration = zeros(3)`: Global frame angular acceleration vector. If time
        varying, this vector may be provided as a function of time.
 """
-function SciMLBase.ODEProblem(system::System, assembly, tspan;
[…]
