### 2022-03-04 ebb9587
test that ODE problem solves
