### 2022-03-07 fe81e9d
initial version of literate.jl documentation
