+++ docs/src/literate/tipmoment.jl
@@ -1,0 +1,264 @@
+# # Cantilever with a Tip Moment
+#
+# This example shows how to predict the behavior of a cantilever beam that is subjected 
+# to a constant tip moment.  This is a common benchmark problem for the geometrically 
+# nonlinear analysis of beams.
+#
+# ![](../assets/cantilever-tipmoment-drawing.svg)
+# 
+#-
+#md # !!! tip
+#md #     This example is also available as a Jupyter notebook:
+#md #     [`tipmoment.ipynb`](@__NBVIEWER_ROOT_URL__/examples/tipmoment.ipynb).
+#-
+
+using GXBeam, LinearAlgebra
+
+L = 12 # inches
+h = w = 1 # inches
+E = 30e6 # lb/in^4 Young's Modulus
+
+A = h*w
+Iyy = w*h^3/12
+Izz = w^3*h/12
+
+## bending moment (applied at end)
+λ = [0.0, 0.4, 0.8, 1.2, 1.6, 1.8, 2.0]
+m = pi*E*Iyy/L
+M = λ*m
+
+## create points
+nelem = 16
+x = range(0, L, length=nelem+1)
+y = zero(x)
+z = zero(x)
+points = [[x[i],y[i],z[i]] for i = 1:length(x)]
+
+## index of endpoints of each beam element
+start = 1:nelem
+stop = 2:nelem+1
+
+## compliance matrix for each beam element
+compliance = fill(Diagonal([1/(E*A), 0, 0, 0, 1/(E*Iyy), 1/(E*Izz)]), nelem)
+
+## create assembly of interconnected nonlinear beams
+assembly = Assembly(points, start, stop, compliance=compliance)
+
+## pre-initialize system storage
+system = System(assembly, true)
+
+## run an analysis for each prescribed bending moment
+states = Vector{AssemblyState{Float64}}(undef, length(M))
+for i = 1:length(M)
+
+    ## create dictionary of prescribed conditions
+    prescribed_conditions = Dict(
+        ## fixed left side
+        1 => PrescribedConditions(ux=0, uy=0, uz=0, theta_x=0, theta_y=0, theta_z=0),
+        ## moment on right side
+        nelem+1 => PrescribedConditions(Mz = M[i])
+    )
+
+    ## perform a static analysis
+    static_analysis!(system, assembly;
+        prescribed_conditions = prescribed_conditions)
+
+    ## post-process the results
+    states[i] = AssemblyState(system, assembly;
+        prescribed_conditions = prescribed_conditions)
+
+end
+
+nothing #hide
+
+# 
+# This problem has a simple analytical solution, which we obtained from "Study of the 
+# Geometric Stiffening Effect: Comparison of Different Formulations" by Juana M. Mayo, 
+# Daniel Garcia-Vallejo, and Jaime Dominguez.
+#
+
+## analytical solution (ρ = E*I/M)
+analytical(x, ρ) = ifelse(ρ == Inf, zeros(3), [ρ*sin(x/ρ)-x, ρ*(1-cos(x/ρ)), 0])
+nothing #hide
+
+#
+# Plotting the results reveals that the analytical and computational results show 
+# excellent agreement.
+# 
+
+using Plots
+#md using Suppressor #hide
+pyplot()
+
+#- 
+
+#md @suppress_err begin #hide
+
+## set up the plot
+plot(
+    xlim = (-0.25, 1.1),
+    xticks = -0.25:0.25:1.0,
+    xlabel = "x/L",
+    ylim = (-0.05, 0.8),
+    yticks = 0.0:0.2:0.8,
+    ylabel = "y/L",
+    aspect_ratio = 1.0,
+    grid = false,
+    overwrite_figure=false
+    )
+
[…]
