+++ docs/src/literate/static-joined-wing.jl
@@ -1,0 +1,373 @@
+# # Static Analysis of a Joined-Wing
+# 
+# In this example we consider the joined-wing model proposed by Blair in "An Equivalent
+# Beam Formulation for Joined-Wings in a Post-Buckled State" and optimized by Green
+# et al. in "Structural Optimization of Joined-Wing Beam Model with Bend-Twist
+# Coupling using Equivalent Static Loads".
+# 
+# ![](../assets/static-joined-wing-drawing.png)
+#
+#-
+#md # !!! tip
+#md #     This example is also available as a Jupyter notebook:
+#md #     [`static-joined-wing.ipynb`](@__NBVIEWER_ROOT_URL__/examples/static-joined-wing.ipynb).
+#-
+#
+
+using GXBeam, LinearAlgebra
+
+## Set endpoints of each beam
+p1 = [-7.1726, -12, -3.21539]
+p2 = [-5.37945, -9, -2.41154]
+p3 = [-3.5863, -6, -1.6077]
+p4 = [-1.79315, -3, -0.803848]
+p5 = [0, 0, 0]
+p6 = [7.1726, -12, 3.21539]
+
+## get transformation matrix for left beams
+
+## transformation from intermediate frame to global frame
+tmp1 = sqrt(p1[1]^2 + p1[2]^2)
+c1, s1 = -p1[1]/tmp1, -p1[2]/tmp1
+rot1 = [c1 -s1 0; s1 c1 0; 0 0 1]
+
+## transformation from local beam frame to intermediate frame
+tmp2 = sqrt(p1[1]^2 + p1[2]^2 + p1[3]^2)
+c2, s2 = tmp1/tmp2, -p1[3]/tmp2
+rot2 = [c2 0 -s2; 0 1 0; s2 0 c2]
+
+Cab_1 = rot1*rot2
+
+## get transformation matrix for right beam
+
+## transformation from intermediate frame to global frame
+tmp1 = sqrt(p6[1]^2 + p6[2]^2)
+c1, s1 = p6[1]/tmp1, p6[2]/tmp1
+rot1 = [c1 -s1 0; s1 c1 0; 0 0 1]
+
+## transformation from local beam frame to intermediate frame
+tmp2 = sqrt(p6[1]^2 + p6[2]^2 + p6[3]^2)
+c2, s2 = tmp1/tmp2, p6[3]/tmp2
+rot2 = [c2 0 -s2; 0 1 0; s2 0 c2]
+
+Cab_2 = rot1*rot2
+
+## beam 1
+L_b1 = norm(p2-p1)
+r_b1 = p1
+nelem_b1 = 5
+lengths_b1, xp_b1, xm_b1, Cab_b1 = discretize_beam(L_b1, r_b1, nelem_b1;
+    frame = Cab_1)
+compliance_b1 = fill(Diagonal([1.05204e-9, 3.19659e-9, 2.13106e-8, 1.15475e-7,
+    1.52885e-7, 7.1672e-9]), nelem_b1)
+
+## beam 2
+L_b2 = norm(p3-p2)
+r_b2 = p2
+nelem_b2 = 5
+lengths_b2, xp_b2, xm_b2, Cab_b2 = discretize_beam(L_b2, r_b2, nelem_b2;
+    frame = Cab_1)
+compliance_b2 = fill(Diagonal([1.24467e-9, 3.77682e-9, 2.51788e-8, 1.90461e-7,
+    2.55034e-7, 1.18646e-8]), nelem_b2)
+
+## beam 3
+L_b3 = norm(p4-p3)
+r_b3 = p3
+nelem_b3 = 5
[…]
