+++ docs/generate.jl
@@ -19,14 +19,9 @@ for example in readdir(EXAMPLEDIR)
         code_clean = strip(code_clean)
 
         mdpost(str) = replace(str, "@__CODE__" => code_clean)
-        function nbpre(str)
-            # \llbracket and \rr bracket not supported by MathJax (Jupyter/nbviewer)
-            str = replace(str, "\\llbracket" => "[\\![", "\\rrbracket" => "]\\!]")
-            return str
-        end
 
         Literate.markdown(input, GENERATEDDIR, postprocess = mdpost)
-        Literate.notebook(input, GENERATEDDIR, preprocess = nbpre, execute = is_ci) # Don't execute locally
+        Literate.notebook(input, GENERATEDDIR, execute = is_ci) # Don't execute locally
     else
         @warn "ignoring $example"
     end
