+++ docs/make.jl
@@ -2,6 +2,9 @@ using Documenter, Literate, GXBeam
 
 const is_ci = haskey(ENV, "GITHUB_ACTIONS")
 
+# Pre-install matplotlib
+import Plots; Plots.pyplot()
+
 # Generate examples
 include("generate.jl")
 
