+++ docs/src/literate/cantilever.jl
@@ -184,10 +184,3 @@ nothing #hide
 
 # Note that we could have easily performed a nonlinear analysis for this problem by setting 
 # `linear=false`.
-
-#md # Here's the complete example, without any comments.  It is also available here:
-#md # [`cantilever.jl`](cantilever.jl).
-#md #
-#md # ```julia
-#md # @__CODE__
-#md # ```
