+++ docs/src/literate/guide.jl
@@ -23,13 +23,13 @@
 import Pkg
 Pkg.add("GXBeam")
 #md end #hide
-nothing #hide
+#!jl nothing #hide
 
 # Now, that the package is installed we need to load it so that we can use it.  It's also 
 # often helpful to load the LinearAlgebra package.
 
 using GXBeam, LinearAlgebra
-nothing #hide
+#!jl nothing #hide
 
 # The geometry we will be working with is a rotating beam with a swept tip as pictured.
 # 
@@ -68,7 +68,7 @@ L_b1 = 31.5 # length of straight section of the beam in inches
 r_b1 = [2.5, 0, 0] # starting point of straight section of the beam
 nelem_b1 = 10 # number of elements in the straight section of the beam
 lengths_b1, xp_b1, xm_b1, Cab_b1 = discretize_beam(L_b1, r_b1, nelem_b1)
-nothing #hide
+#!jl nothing #hide
 
 # The length of each beam element is equal since we used the number of elements to define 
 # the discretization.  Alternatively we can manually specify the discretization of the 
@@ -80,7 +80,7 @@ disc_b1 = range(0, 1, length=nelem_b1+1)
 ## discretize straight beam section
 lengths_b1, xp_b1, xm_b1, Cab_b1 = discretize_beam(L_b1, r_b1, disc_b1)
 
-nothing #hide
+#!jl nothing #hide
 
 # We now create the geometry for the swept portion of the wing.  To do so we use the same 
 # [`discretize_beam`](@ref) function, but use the additional keyword argument `frame` in 
@@ -110,7 +110,7 @@ e3 = [0, 0, 1] # axis 3
 frame_b2 = hcat(e1, e2, e3) # transformation matrix from local to body frame
 lengths_b2, xp_b2, xm_b2, Cab_b2 = discretize_beam(L_b2, r_b2, nelem_b2;
     frame = frame_b2)
-nothing #hide
+#!jl nothing #hide
 
 # We will now manually combine the results of our two calls to `discretize_beam`.  Since 
 # the last endpoint from the straight section is the same as the first endpoint of the 
@@ -125,7 +125,7 @@ lengths = vcat(lengths_b1, lengths_b2) # length of each beam element in our asse
 midpoints = vcat(xm_b1, xm_b2) # midpoint of each beam element in our assembly
 Cab = vcat(Cab_b1, Cab_b2) # transformation matrix from local to body frame
                            ## for each beam element in our assembly
-nothing #hide
+#!jl nothing #hide
 
 # Next we need to define the stiffness (or compliance) and mass matrices for each 
 # beam element.
@@ -249,7 +249,7 @@ compliance = fill(Diagonal([1/(E*A), 1/(G*Ay), 1/(G*Az), 1/(G*Jx), 1/(E*Iyy),
     1/(E*Izz)]), nelem)
 
 mass = fill(Diagonal([ρ*A, ρ*A, ρ*A, ρ*J, ρ*Iyy, ρ*Izz]), nelem)
-nothing #hide
+#!jl nothing #hide
 
 # Our case is simple enough that we can analytically calculate most values for the 
 # compliance and mass matrices, but this is not generally the case.  For more complex 
@@ -272,7 +272,7 @@ assembly = Assembly(points, start, stop;
    frames = Cab,
    lengths = lengths,
    midpoints = midpoints)
-nothing #hide
+#!jl nothing #hide
 
[…]
