### 2022-03-10 #3997 @taylormcd
@Kevin-Mattheus-Moerman
Jupyter notebook versions of the examples are now available.
