### 2022-03-10 f335e2f
don't explicitly installpackage when running code for guide
