### 2022-03-10 75c2eca
move is_ci flag to generate.jl
