### 2022-03-13 #3997 @Kevin-Mattheus-Moerman
@taylormcd thanks for adding those. I'll have a look later the coming week. Where can I find the new Jupyter notebook versions?
