### 2022-03-14 #3997 @taylormcd
They're linked from the relevant pages in the documentation. Note that I haven't yet created a new release with the changes, so the notebooks are only linked from the development version of the documentation..
