### 2022-03-14 caf8b84
allow for massless elements during time domain initialization; fixes #71
