### 2022-03-17 53567a3
add flag for enabling/disabling structural damping
