+++ src/element.jl
@@ -628,6 +628,7 @@ condition analysis.
  - `elem`: beam element
  - `distributed_loads`: dictionary with all distributed loads
  - `point_masses`: dictionary with all point masses
+ - `structural_damping`: flag indicating whether to apply structural damping
  - `gvec`: gravity vector
  - `force_scaling`: scaling parameter for forces
  - `icol`: starting index for the beam's state variables
@@ -652,8 +653,8 @@ condition analysis.
  - `Mdot0`: initial resultant moment for the beam element 
 """
 @inline function initial_condition_element_residual!(resid, x, ielem, elem, 
-    distributed_loads, point_masses, gvec, force_scaling, icol, irow_e, irow_e1, irow_p1, 
-    irow_e2, irow_p2, x0, v0, ω0, a0, α0, u, θ, udot, θdot, Fdot, Mdot)
+    distributed_loads, point_masses, structural_damping, gvec, force_scaling, icol, irow_e, 
+    irow_e1, irow_p1, irow_e2, irow_p2, x0, v0, ω0, a0, α0, u, θ, udot, θdot, Fdot, Mdot)
 
     # element properties
     ΔL = elem.L
@@ -677,8 +678,9 @@ condition analysis.
         Ωdot = @SVector zeros(3)       
 
         return dynamic_element_residual!(resid, x, ielem, elem, distributed_loads, 
-            point_masses, gvec, force_scaling, icol, irow_e, irow_e1, irow_p1, irow_e2, 
-            irow_p2, x0, v0, ω0, a0, α0, udot, θdot, Vdot, Ωdot)
+            point_masses, structural_damping, gvec, force_scaling, 
+            icol, irow_e, irow_e1, irow_p1, irow_e2, irow_p2, 
+            x0, v0, ω0, a0, α0, udot, θdot, Fdot, Mdot, Vdot, Ωdot)
     end
 
     # element state variables
@@ -746,13 +748,11 @@ condition analysis.
         m2 += dload.m2 + Ct*dload.m2_follower
     end
 
-    # damping loads
-    Fd = SVector(μ[1]*Fdot[1], μ[2]*Fdot[2], μ[3]*Fdot[3])
-    Md = SVector(μ[4]*Mdot[1], μ[5]*Mdot[2], μ[6]*Mdot[3])
-
-    # total element loads
-    F = F + Fd
-    M = M + Md
+    # add structural damping
+    if structural_damping
+        F += SVector(μ[1]*Fdot[1], μ[2]*Fdot[2], μ[3]*Fdot[3])
+        M += SVector(μ[4]*Mdot[1], μ[5]*Mdot[2], μ[6]*Mdot[3])
+    end
 
     # solve for the element resultants
     f_u1, f_u2, f_ψ1, f_ψ2, f_F1, f_F2, f_M1, f_M2, f_V, f_Ω =
@@ -781,6 +781,7 @@ time-marching analysis.
  - `elem`: beam element
  - `distributed_loads`: dictionary with all distributed loads
  - `point_masses`: dictionary with all point masses
+ - `structural_damping`: flag indicating whether to apply structural damping
  - `gvec`: gravity vector
  - `force_scaling`: scaling parameter for forces
  - `icol`: starting index for the beam's state variables
@@ -805,9 +806,10 @@ time-marching analysis.
  - `Ωdot_init`: `2/dt*Ω + Ωdot` for the beam element from the previous time step
  - `dt`: time step size
 """
[…]

[…]
