### 2022-03-17 9d50632
save image to correct location for docs
