+++ src/section.jl
@@ -1,0 +1,487 @@
+using LinearAlgebra: I, Symmetric, det
+
+struct Material{TF}
+    E1::TF
+    E2::TF
+    E3::TF
+    nu12::TF
+    nu13::TF
+    nu23::TF
+    G12::TF
+    G13::TF
+    G23::TF
+    rho::TF
+end
+
+Material(E1, E2, nu12, G12, rho) = Material(E1, E2, E2, nu12, nu12, nu12, G12, G12, G12, rho)
+
+struct Node{TF, TI}
+    x::TF
+    y::TF
+    number::TI
+    material::Material
+    theta::TF
+    beta::TF
+end
+
+function stiffness(material) 
+    E1 = material.E1; E2 = material.E2; E3 = material.E3
+    nu12 = material.nu12; nu13 = material.nu13; nu23 = material.nu23
+    G12 = material.G12; G13 = material.G13; G23 = material.G23
+
+    nu21 = nu12*E2/E1
+    nu31 = nu13*E3/E1
+    nu32 = nu23*E3/E2
+    delta = 1.0 / (1 - nu12*nu21 - nu23*nu32 - nu13*nu31 - 2*nu21*nu32*nu13)
+
+    Q = zeros(eltype(E1), 6, 6)
+    Q[6, 6] = E1*(1 - nu23*nu32)*delta
+    Q[1, 1] = E2*(1 - nu13*nu31)*delta
+    Q[2, 2] = E3*(1 - nu12*nu21)*delta
+    Q[1, 6] = E1*(nu21 + nu31*nu23)*delta
+    Q[2, 6] = E1*(nu31 + nu21*nu32)*delta
+    Q[1, 2] = E2*(nu32 + nu12*nu31)*delta
+    Q[4, 4] = G12
+    Q[5, 5] = G13
+    Q[3, 3] = G23
+
+    return Symmetric(Q)
+end
+
+function rotate_to_ply(K, theta)
+    c = cos(theta)
+    s = sin(theta)
+    T = [s^2 c^2 2*s*c 0 0 0;
+        0.0 0 0 0 0 1;
+        0 0 0 s c 0;
+        s*c -s*c c^2-s^2 0 0 0;
+        0 0 0 c -s 0;
+        c^2 s^2 -2*s*c 0 0 0]
+
+    return T*K*T'
+end
+
+function rotate_to_element(K, beta)
+    c = cos(beta)
+    s = sin(beta)
+    T = [c^2 s^2 -2*s*c 0 0 0;
+          s^2 c^2 2*s*c 0 0 0;
+          s*c -s*c c^2-s^2 0 0 0;
+          0 0 0 c -s 0;
+          0 0 0 s c 0;
+          0.0 0 0 0 0 1]
+
+    return T*K*T'
+end
+
+# function rotate_to_element(K, x, y)
+#     dx = x[2] - x[1]
+#     dy = y[2] - y[1]
+#     ds = sqrt(dx^2 + dy^2)
+#     c = dx/ds
+#     s = dy/ds
+#     T = [c^2 s^2 0 -2*s*c 0 0;
+#           s^2 c^2 0 2*s*c 0 0;
+#           0.0 0 1 0 0 0;
+#           s*c -s*c 0 c^2-s^2 0 0;
+#           0 0 0 0 c -s;
+#           0 0 0 0 s c]
+
+#     return T*K*T'
+# end
+
+
+# function transformation(Q, R)
+#     Qnew = zeros(6, 6)
+#     for i = 1:3
+#         for j = i:3
+#             for k = 1:3
+#                 for l = 1:3
+#                     Qnew[i, j] = R[i, k]*R[j, l]*Q[k, l]
+#                 end
+#             end
[…]
