### 2022-03-23 79e0162
modify default force scaling
