### 2022-03-23 908c878
found error in stiffness matrix computation
