### 2022-03-24 b065d00
redefine input structs to material is associated with element rather than nodes, needed to  pass case 2 consistently
