+++ src/section.jl
@@ -19,6 +19,10 @@ struct Node{TF, TI}
     x::TF
     y::TF
     number::TI
+end
+
+struct Element{VI, TF}
+    nodenum::VI
     material::Material
     theta::TF
     beta::TF
@@ -131,7 +135,7 @@ function elementQ(material, theta, beta)
 end
 
 
-function elementintegrand(ksi, eta, nodes)
+function elementintegrand(ksi, eta, element, nodes)
 
     # shape functions
     N = zeros(4)
@@ -151,11 +155,7 @@ function elementintegrand(ksi, eta, nodes)
     # basic matrices
     Z = [I [0.0 0 -y; 0 0 x; y -x 0]]  # translation and cross product
     S = [zeros(3, 3); I]
-    Q = zeros(6, 6)
-    for i = 1:4  # TODO: is this the right thing to do?
-        Q += elementQ(nodes[i].material, nodes[i].theta, nodes[i].beta)
-    end
-    Q /= 4
+    Q = elementQ(element.material, element.theta, element.beta)
     N = [N[1]*Matrix(1.0I, 3, 3) N[2]*I N[3]*I N[4]*I]
     SZ = S*Z
     SN = S*N
@@ -214,7 +214,8 @@ function elementintegrand(ksi, eta, nodes)
     A = SZ'*Q*SZ*detJ
     R = BN'*Q*SZ*detJ
     E = BN'*Q*BN*detJ
-    C = SN'*Q*BN*detJ  # NOTE: error in implementation sectiono of manual, theory is correct
+    # C = SN'*Q*BN*detJ  # NOTE: error in implementation sectiono of manual, theory is correct
+    C = BN'*Q*SN*detJ  # use Giavottoa definition
     L = SN'*Q*SZ*detJ
     M = SN'*Q*SN*detJ
 
@@ -222,15 +223,15 @@ function elementintegrand(ksi, eta, nodes)
 end
 
 
-function submatrix(nodes)
+function submatrix(element, nodes)
 
     # 2-point Gauss quadrature in both dimensions
     ksi = [1.0/sqrt(3), 1.0/sqrt(3), -1.0/sqrt(3), -1.0/sqrt(3)]
     eta = [1.0/sqrt(3), -1.0/sqrt(3), 1.0/sqrt(3), -1.0/sqrt(3)]
 
-    A, R, E, C, L, M = elementintegrand(ksi[1], eta[1], nodes)
+    A, R, E, C, L, M = elementintegrand(ksi[1], eta[1], element, nodes)
     for i = 2:4
-        Ai, Ri, Ei, Ci, Li, Mi = elementintegrand(ksi[i], eta[i], nodes)
+        Ai, Ri, Ei, Ci, Li, Mi = elementintegrand(ksi[i], eta[i], element, nodes)
         A += Ai; R += Ri; E += Ei
         C += Ci; L += Li; M += Mi
     end
@@ -248,15 +249,15 @@ function node2idx(nodes)
     return idx
 end
 
-function compliancematrix(nodes, connectivity)
+function sectionprops(nodes, elements)
 
     # initialize
-    ne, _ = size(connectivity) # number of elements
+    ne = length(elements) # number of elements
     nn = length(nodes)  # number of nodes
     ndof = 3 * nn  # 3 displacement dof per node
     
     # initialize
-    etype = eltype(nodes[1].theta)  # TODO add others
+    etype = eltype(elements[1].theta)  # TODO add others
     A = zeros(etype, 6, 6)  # 6 x 6
[…]
