### 2022-03-30 300745a
save matrix factorization, use sparse matrices, compute beta internally.  a couple more unit tests
