+++ src/section.jl
@@ -1,4 +1,5 @@
-using LinearAlgebra: I, Symmetric, det
+using LinearAlgebra: I, Symmetric, det, factorize
+using SparseArrays: spzeros, sparse
 
 struct Material{TF}
     E1::TF
@@ -25,7 +26,6 @@ struct Element{VI, TF}
     nodenum::VI
     material::Material
     theta::TF
-    beta::TF
 end
 
 function stiffness(material) 
@@ -65,9 +65,7 @@ function rotate_to_ply(K, theta)
     return T*K*T'
 end
 
-function rotate_to_element(K, beta)
-    c = cos(beta)
-    s = sin(beta)
+function rotate_to_element(K, c, s)  # c = cos(beta), s = sin(beta)
     T = [c^2 s^2 -2*s*c 0 0 0;
           s^2 c^2 2*s*c 0 0 0;
           s*c -s*c c^2-s^2 0 0 0;
@@ -126,10 +124,10 @@ end
 #     return Q[idx, idx]
 # end
 
-function elementQ(material, theta, beta)
+function elementQ(material, theta, cbeta, sbeta)
     Q1 = stiffness(material)  # material
     Q2 = rotate_to_ply(Q1, theta)
-    Q3 = rotate_to_element(Q2, beta)
+    Q3 = rotate_to_element(Q2, cbeta, sbeta)
 
     return Q3
 end
@@ -151,11 +149,22 @@ function elementintegrand(ksi, eta, element, nodes)
         x += N[i]*nodes[i].x
         y += N[i]*nodes[i].y
     end
+
+    # orientation (beta)
+    xl = 0.5*(nodes[1].x + nodes[4].x)
+    yl = 0.5*(nodes[1].y + nodes[4].y)
+    xr = 0.5*(nodes[2].x + nodes[3].x)
+    yr = 0.5*(nodes[2].y + nodes[3].y)
+    dx = xr - xl
+    dy = yr - yl
+    ds = sqrt(dx^2 + dy^2)
+    cbeta = dx/ds
+    sbeta = dy/ds
     
     # basic matrices
     Z = [I [0.0 0 -y; 0 0 x; y -x 0]]  # translation and cross product
     S = [zeros(3, 3); I]
-    Q = elementQ(element.material, element.theta, element.beta)
+    Q = elementQ(element.material, element.theta, cbeta, sbeta)
     N = [N[1]*Matrix(1.0I, 3, 3) N[2]*I N[3]*I N[4]*I]
     SZ = S*Z
     SN = S*N
@@ -256,14 +265,16 @@ function sectionprops(nodes, elements)
     nn = length(nodes)  # number of nodes
     ndof = 3 * nn  # 3 displacement dof per node
     
+    # TODO: benchmark sparsity
+
     # initialize
     etype = eltype(elements[1].theta)  # TODO add others
     A = zeros(etype, 6, 6)  # 6 x 6
     R = zeros(etype, ndof, 6)  #  nn*3 x 6
-    E = zeros(etype, ndof, ndof)  # nn*3 x nn*3
-    C = zeros(etype, ndof, ndof)  # nn*3 x nn*3
+    E = spzeros(etype, ndof, ndof)  # nn*3 x nn*3
+    C = spzeros(etype, ndof, ndof)  # nn*3 x nn*3
     L = zeros(etype, ndof, 6)  # nn*3 x 6
-    M = zeros(etype, ndof, ndof)  # nn*3 x nn*3
+    M = spzeros(etype, ndof, ndof)  # nn*3 x nn*3
 
[…]
