### 2022-03-30 bbc8607
working through test cases
