+++ src/postprocess.jl
@@ -4,16 +4,20 @@
 Holds the state variables for a point
 
 # Fields:
- - `u`: Displacement variables for the point (in the global coordinate frame)
- - `theta`: Wiener-Milenkovic rotational displacement variables for the point
- - `F`: Externally applied forces on the point (in the global coordinate frame)
- - `M`: Externally applied moments on the point (in the global coordinate frame)
+ - `F`: Externally applied forces on the point
+ - `M`: Externally applied moments on the point
+ - `u`: Displacement variables for the point
+ - `theta`: Displacement Wiener-Milenkovic rotation parameters for the point
+ - `V`: Linear velocity of the point
+ - `Ω`: Angular velocity of the point
 """
 struct PointState{TF}
     u::SVector{3, TF}
     theta::SVector{3, TF}
     F::SVector{3, TF}
     M::SVector{3, TF}
+    V::SVector{3, TF}
+    Omega::SVector{3, TF}
 end
 
 """
@@ -22,9 +26,8 @@ end
 Holds the state variables for an element
 
 # Fields:
- - `u`: Displacement variables for the element (in the global coordinate frame)
- - theta: Wiener-Milenkovic rotational displacement variables for the element
-       (in the global coordinate frame)
+ - `u`: Displacement variables for the element
+ - `theta`: Wiener-Milenkovic rotational displacement variables for the element
  - `F`: Resultant forces for the element (in the deformed local coordinate frame)
  - `M`: Resultant moments for the element (in the deformed local coordinate frame)
  - `V`: Linear velocity of the element (in the deformed local coordinate frame)
@@ -71,7 +74,7 @@ function AssemblyState(system, assembly, x = system.x;
 
     points = extract_point_states(system, assembly, x; prescribed_conditions)
 
-    elements = extract_element_states(system, x)
+    elements = extract_element_states(system, assembly, x; prescribed_conditions)
 
     return AssemblyState(points, elements)
 end
@@ -90,64 +93,23 @@ analysis.
 function extract_point_state(system, assembly, ipoint, x = system.x;
     prescribed_conditions = Dict{Int,PrescribedConditions{Float64}}())
 
-    pc = typeof(prescribed_conditions) <: AbstractDict ? prescribed_conditions :
-        prescribed_conditions(system.t)
+    @unpack force_scaling, dynamic_indices, t = system
 
-    force_scaling = system.force_scaling
+    pc = typeof(prescribed_conditions) <: AbstractDict ? prescribed_conditions : prescribed_conditions(t)
 
-    icol_p = system.icol_point[ipoint]
-
-    if icol_p <= 0
-        # point variables are not state variables, solve for point variables
-
-        # find the first beam connected to the point, checking both sides
-        ielem = findfirst(x -> x == ipoint, assembly.start)
-        if !isnothing(ielem)
-            side = -1
-        else
-            ielem = findfirst(x -> x == ipoint, assembly.stop)
-            side = 1
-        end
-
-        elem = assembly.elements[ielem]
-
-        # get beam state variables
-        icol_b = system.icol_elem[ielem]
-        u_e = SVector(x[icol_b   ], x[icol_b+1 ], x[icol_b+2 ])
-        θ_b = SVector(x[icol_b+3 ], x[icol_b+4 ], x[icol_b+5 ])
-        F_e = SVector(x[icol_b+6 ], x[icol_b+7 ], x[icol_b+8 ]) .* force_scaling
-        M_e = SVector(x[icol_b+9 ], x[icol_b+10], x[icol_b+11]) .* force_scaling
-
-        # get beam element properties
-        ΔL_b = elem.L
[…]
