+++ src/element.jl
@@ -91,2896 +91,1700 @@ the element expressed in the deformed beam element frame
 end
 
 """
-    static_element_state_variables(x, icol, force_scaling)
+    point_displacement(x, ielem, icol_elem, force_scaling)
 
-Return element state variables for a static analysis
+Extract the state variables `F` and `M` of element `ielem` from the state variable vector.
 """
-@inline function static_element_state_variables(x, icol, force_scaling)
-    u = SVector(x[icol   ], x[icol+1 ], x[icol+2 ])
-    θ = SVector(x[icol+3 ], x[icol+4 ], x[icol+5 ])
-    F = SVector(x[icol+6 ], x[icol+7 ], x[icol+8 ]) .* force_scaling
-    M = SVector(x[icol+9 ], x[icol+10], x[icol+11]) .* force_scaling
-    return u, θ, F, M
+@inline function element_states(x, ielem, icol_elem, force_scaling)
+
+    icol = icol_elem[ielem]
+
+    F = SVector(x[icol  ], x[icol+1], x[icol+2]) .* force_scaling
+    M = SVector(x[icol+3], x[icol+4], x[icol+5]) .* force_scaling
+
+    return F, M
 end
 
 """
-    dynamic_element_state_variables(x, icol, force_scaling)
+    static_element_properties(x, indices, force_scaling, assembly, ielem, 
+        prescribed_conditions, gravity)
 
-Return element state variables for a dynamic analysis
+Calculate/extract the element properties needed to construct the residual for a static 
+analysis
 """
-@inline function dynamic_element_state_variables(x, icol, force_scaling)
-    u = SVector(x[icol   ], x[icol+1 ], x[icol+2 ]) # or Vdot
-    θ = SVector(x[icol+3 ], x[icol+4 ], x[icol+5 ]) # or Ωdot
-    F = SVector(x[icol+6 ], x[icol+7 ], x[icol+8 ]) .* force_scaling
-    M = SVector(x[icol+9 ], x[icol+10], x[icol+11]) .* force_scaling
-    V = SVector(x[icol+12], x[icol+13], x[icol+14])
-    Ω = SVector(x[icol+15], x[icol+16], x[icol+17])
-    return u, θ, F, M, V, Ω
+@inline function static_element_properties(x, indices, force_scaling, assembly, ielem, 
+    prescribed_conditions, gravity)
+
+    # unpack element parameters
+    @unpack L, Cab, compliance, mass = assembly.elements[ielem]
+
+    # compliance submatrices
+    S11 = compliance[SVector{3}(1:3), SVector{3}(1:3)]
+    S12 = compliance[SVector{3}(1:3), SVector{3}(4:6)]
+    S21 = compliance[SVector{3}(4:6), SVector{3}(1:3)]
+    S22 = compliance[SVector{3}(4:6), SVector{3}(4:6)]
+
+    # mass submatrices
+    mass11 = mass[SVector{3}(1:3), SVector{3}(1:3)]
+    mass12 = mass[SVector{3}(1:3), SVector{3}(4:6)]
+    mass21 = mass[SVector{3}(4:6), SVector{3}(1:3)]
+    mass22 = mass[SVector{3}(4:6), SVector{3}(4:6)]
+
+    # starting node displacement
+    u1, θ1 = point_displacement(x, assembly.start[ielem], indices.icol_point, prescribed_conditions)
+
+    # ending node displacement
+    u2, θ2 = point_displacement(x, assembly.stop[ielem], indices.icol_point, prescribed_conditions)
+
+    # beam element displacement
[…]
