+++ src/section.jl
@@ -76,53 +76,11 @@ function rotate_to_element(K, c, s)  # c = cos(beta), s = sin(beta)
     return T*K*T'
 end
 
-# function rotate_to_element(K, x, y)
-#     dx = x[2] - x[1]
-#     dy = y[2] - y[1]
-#     ds = sqrt(dx^2 + dy^2)
-#     c = dx/ds
-#     s = dy/ds
-#     T = [c^2 s^2 0 -2*s*c 0 0;
-#           s^2 c^2 0 2*s*c 0 0;
-#           0.0 0 1 0 0 0;
-#           s*c -s*c 0 c^2-s^2 0 0;
-#           0 0 0 0 c -s;
-#           0 0 0 0 s c]
 
-#     return T*K*T'
-# end
-
-
-# function transformation(Q, R)
-#     Qnew = zeros(6, 6)
-#     for i = 1:3
-#         for j = i:3
-#             for k = 1:3
-#                 for l = 1:3
-#                     Qnew[i, j] = R[i, k]*R[j, l]*Q[k, l]
-#                 end
-#             end
-#         end
-#     end
-#     return Symmetric(Qnew)
-# end
-
-
-# Rmaterial(theta) = [cos(theta) 0.0 sin(theta); 0 1 0; -sin(theta) 0 cos(theta)]
-# Rgeometry(beta) = [cos(beta) -sin(beta) 0.0; sin(beta) cos(beta) 0; 0 0 1]
-# function Rgeometry(y1, y2, z2, z2)
-#     dy = y2 - y1
-#     dz = z2 - z1
-#     ds = sqrt(dy^2 + dz^2)
-#     cb = dx/ds
-#     sb = dy/ds
-#     return [cb -sb 0.0; sb cb 0; 0 0 1]
-# end
-
-# function reorder(Q)
-#     idx = [2, 3, 6, 4, 5, 1]
-#     return Q[idx, idx]
-# end
+function reorder(K)  # reorder to GXBeam format
+    idx = [3, 1, 2, 6, 4, 5]
+    return K[idx, idx]
+end
 
 function elementQ(material, theta, cbeta, sbeta)
     Q1 = stiffness(material)  # material
@@ -700,150 +658,256 @@ xs, ys, xt, yt = centers(S)
 @test isapprox(xt, -6.051E-02, atol=0.002e-2) 
 @test isapprox(yt, 0.0, atol=1e-6) 
 
-# # A critical assessment of computer tools for calculating composite wind turbine blade properties
 
-# # ----- multi-layer composite pipe --------
+# A critical assessment of computer tools for calculating composite wind turbine blade properties
+# ------ circular tube -------
+E = 73e9
+nu = 0.33
+G = E/(2*(1 + nu))
+tratio = 1.0/3
+R = 0.3
+t = tratio*2*R
+circmat = Material(E, E, nu, G, 1.0)
+
+nr = 20
+nt = 100
+r = range(R - t, R, length=nr)
+theta = range(0.0, 2*pi, length=nt)
+
+nodes = Vector{Node}(undef, nr*(nt-1))
+elements = Vector{Element}(undef, (nr-1)*(nt-1))
+let 
+m = 1
+for i = 1:nt-1
+    for j = 1:nr
+        nodes[m] = Node(r[j]*cos(theta[i]), r[j]*sin(theta[i]), m)
+        m += 1
+    end
+end
+
+n = 1
+for i = 1:nt-1
+    for j = 1:nr-1
+        if i == nt-1
+            ip = 0
+        else
+            ip = i
+        end
+
+        elements[n] = Element([nr*ip+j, nr*(i-1)+j, nr*(i-1)+j+1, nr*ip+j+1], circmat, 0.0)
+        n += 1
+    end
+end
+end
+
[…]
