+++ src/point.jl
@@ -1,47 +1,4 @@
 """
-    point_displacement(x, ipoint, icol_point, prescribed_conditions)
-
-Extract the displacements `u` and `θ` of point `ipoint` from the state variable vector or 
-prescribed conditions.
-"""
-@inline function point_displacement(x, ipoint, icol_point, prescribed_conditions)
-
-    if haskey(prescribed_conditions, ipoint)
-        u, θ = point_displacement(x, icol_point[ipoint], prescribed_conditions[ipoint])
-    else
-        u, θ = point_displacement(x, icol_point[ipoint])
-    end
-
-    return u, θ
-end
-
-@inline function point_displacement(x, icol, prescribed_conditions)
-
-    # unpack prescribed conditions for the node
-    @unpack value, follower, isforce = prescribed_conditions
-
-    # node linear displacement
-    u = SVector(ifelse(isforce[1], x[icol  ], value[1]),
-                ifelse(isforce[2], x[icol+1], value[2]),
-                ifelse(isforce[3], x[icol+2], value[3]))
-
-    # node angular displacement
-    θ = SVector(ifelse(isforce[4], x[icol+3], value[4]),
-               ifelse(isforce[5], x[icol+4], value[5]),
-               ifelse(isforce[6], x[icol+5], value[6]))
-
-    return u, θ
-end
-
-@inline function point_displacement(x, icol)
-
-    u = SVector(x[icol], x[icol+1], x[icol+2])
-    θ = SVector(x[icol+3], x[icol+4], x[icol+5])
-
-    return u, θ
-end
-
-"""
     point_loads(x, ipoint, icol, force_scaling, prescribed_conditions)
 
 Extract the loads `F` and `M` of point `ipoint` from the state variable vector or 
@@ -110,149 +67,6 @@ end
     M = @SVector zeros(eltype(x), 3)
 
     return F, M
-end
-
-"""
-    point_velocities(x, ipoint, icol_point)
-
-Extract the velocities `V` and `Ω` of point `ipoint` from the state variable vector
-"""
-@inline function point_velocities(x, ipoint, icol_point)
-
-    V, Ω = point_velocities(x, icol_point[ipoint])
-
-    return V, Ω
-end
-
-@inline function point_velocities(x, icol)
-
-    V = SVector(x[icol+6], x[icol+7], x[icol+8])
-    Ω = SVector(x[icol+9], x[icol+10], x[icol+11])
-
-    return V, Ω
-end
-
-"""
-    point_displacement_rates(dx, ipoint, icol, prescribed_conditions)
-
-Extract the displacement rates `udot` and `θdot` of point `ipoint` from the rate variable vector.
-"""
-@inline function point_displacement_rates(dx, ipoint, icol, prescribed_conditions)
-
-    if haskey(prescribed_conditions, ipoint)
-        udot, θdot = point_displacement_rates(dx, icol[ipoint], prescribed_conditions[ipoint])
-    else
-        udot, θdot = point_displacement_rates(dx, icol[ipoint])
-    end
-
-    return udot, θdot
-end
-
-@inline function point_displacement_rates(dx, icol, prescribed_conditions)
-
-    # unpack prescribed conditions for the node
-    @unpack value, follower, isforce = prescribed_conditions
-
-    # node linear displacement
-    udot = SVector(ifelse(isforce[1], dx[icol  ], zero(eltype(dx))),
-                   ifelse(isforce[2], dx[icol+1], zero(eltype(dx))),
[…]
