### 2022-04-06 byuflowlab/GXBeam.jl#77 @Kevin-Mattheus-Moerman
Tests failing
Uppon running `] test GXBeam` I get the following:

```julia
     Testing Running tests...
Test Summary: | Pass  Total
Math          |   15     15
Test Summary:                         | Pass  Total
Jacobian and Mass Matrix Calculations |    6      6
Test Summary:                                                              | Pass  Total
Linear Analysis of a Cantilever Partially Under a Uniform Distributed Load |   62     62
Test Summary:                                             | Pass  Total
Linear Analysis of a Beam Under a Linear Distributed Load |   82     82
Test Summary:                                                       | Pass  Total
Nonlinear Analysis of a Cantilever Subjected to a Constant Tip Load |   99     99
Test Summary:                                                     | Pass  Total
Nonlinear Analysis of a Cantilever Subjected to a Constant Moment |  462    462
Test Summary:                                                  | Pass  Total
Nonlinear Analysis of the Bending of a Curved Beam in 3D Space |    3      3
Test Summary:                  | Pass  Total
Rotating Beam with a Swept Tip |   48     48
Test Summary:                                      | Pass  Total
Nonlinear Dynamic Analysis of a Wind Turbine Blade |    1      1
Test Summary:                              | Pass  Total
Nonlinear Static Analysis of a Joined-Wing |  423    423
Test Summary:                               | Pass  Total
Nonlinear Dynamic Analysis of a Joined-Wing |    1      1
┌ Warning: dt <= dtmin. Aborting. There is either an error in your model specification or the true solution is unstable.
└ @ SciMLBase ~/.julia/packages/SciMLBase/GW7GW/src/integrator_interface.jl:345
DifferentialEquations: Test Failed at /home/kevin/.julia/packages/GXBeam/S8Yc6/test/runtests.jl:1200
  Expression: sol.t[end] == 2.0
   Evaluated: 0.47166509384056277 == 2.0
Stacktrace:
 [1] macro expansion
   @ ~/julia/share/julia/stdlib/v1.7/Test/src/Test.jl:445 [inlined]
 [2] macro expansion
   @ ~/.julia/packages/GXBeam/S8Yc6/test/runtests.jl:1200 [inlined]
 [3] macro expansion
   @ ~/julia/share/julia/stdlib/v1.7/Test/src/Test.jl:1283 [inlined]
 [4] top-level scope
   @ ~/.julia/packages/GXBeam/S8Yc6/test/runtests.jl:1141
DifferentialEquations: Error During Test at /home/kevin/.julia/packages/GXBeam/S8Yc6/test/runtests.jl:1139
  Got exception outside of a @test
  type Nothing has no field α
  Stacktrace:
    [1] setproperty!(x::Nothing, f::Symbol, v::Float64)
      @ Base ./Base.jl:43
    [2] calc_W!
      @ ~/.julia/packages/OrdinaryDiffEq/iOlmD/src/derivative_utils.jl:559 [inlined]
    [3] update_W!
      @ ~/.julia/packages/OrdinaryDiffEq/iOlmD/src/derivative_utils.jl:663 [inlined]
    [4] nlsolve!(nlsolver::OrdinaryDiffEq.NLSolver{NLNewton{Rational{Int64}, Rational{Int64}, Rational{Int64}}, true, Vector{Float64}, Rational{Int64}, Nothing, Float64, OrdinaryDiffEq.NLNewtonCache{Vector{Float64}, Float64, Float64, Vector{Float64}, Matrix{Float64}, Matrix{Float64}, Nothing, Nothing, LinearSolve.LinearCache{Matrix{Float64}, Vector{Float64}, Vector{Float64}, SciMLBase.NullParameters, Nothing, LU{Float64, Matrix{Float64}}, LinearSolve.InvPreconditioner{Diagonal{Float64, Vector{Float64}}}, Diagonal{Float64, Vector{Float64}}, Float64}}}, integrator::OrdinaryDiffEq.ODEIntegrator{DABDF2{12, true, Nothing, NLNewton{Rational{Int64}, Rational{Int64}, Rational{Int64}}, typeof(OrdinaryDiffEq.DEFAULT_PRECS), Val{:forward}, true, nothing}, true, Vector{Float64}, Vector{Float64}, Float64, Tuple{var"#103#105"{Int64}, Dict{Int64, DistributedLoads{Float64}}, Dict{Int64, Vector{PointMass{Float64}}}, StaticArrays.SVector{3, Float64}, StaticArrays.SVector{3, Float64}, StaticArrays.SVector{3, Float64}, StaticArrays.SVector{3, Float64}, StaticArrays.SVector{3, Float64}, StaticArrays.SVector{3, Float64}}, Float64, Float64, Float64, Float64, Vector{Vector{Float64}}, DAESolution{Float64, 2, Vector{Vector{Float64}}, Nothing, Nothing, Nothing, Vector{Float64}, DAEProblem{Vector{Float64}, Vector{Float64}, Tuple{Float64, Float64}, true, Tuple{var"#103#105"{Int64}, Dict{Int64, DistributedLoads{Float64}}, Dict{Int64, Vector{PointMass{Float64}}}, StaticArrays.SVector{3, Float64}, StaticArrays.SVector{3, Float64}, StaticArrays.SVector{3, Float64}, StaticArrays.SVector{3, Float64}, StaticArrays.SVector{3, Float64}, StaticArrays.SVector{3, Float64}}, DAEFunction{true, GXBeam.var"#299#301"{Assembly{Float64, Vector{Vector{Float64}}, UnitRange{Int64}, Vector{GXBeam.Element{Float64}}}, Float64, Vector{Int64}, Vector{Int64}, Vector{Int64}, Vector{Int64}, Vector{Int64}, Vector{Int64}}, Nothing, Nothing, GXBeam.var"#300#302"{Assembly{Float64, Vector{Vector{Float64}}, UnitRange{Int64}, Vector{GXBeam.Element{Float64}}}, Float64, Vector{Int64}, Vector{Int64}, Vector{Int64}, Vector{Int64}, Vector{Int64}, Vector{Int64}}, Nothing, Nothing, Nothing, SparseArrays.SparseMatrixCSC{Bool, Int64}, Nothing, Nothing, Nothing, Nothing, typeof(SciMLBase.DEFAULT_OBSERVED), Nothing}, Base.Pairs{Symbol, Union{}, Tuple{}, NamedTuple{(), Tuple{}}}, Vector{Bool}}, DABDF2{12, true, Nothing, NLNewton{Rational{Int64}, Rational{Int64}, Rational{Int64}}, typeof(OrdinaryDiffEq.DEFAULT_PRECS), Val{:forward}, true, nothing}, OrdinaryDiffEq.InterpolationData{DAEFunction{true, GXBeam.var"#299#301"{Assembly{Float64, Vector{Vector{Float64}}, UnitRange{Int64}, Vector{GXBeam.Element{Float64}}}, Float64, Vector{Int64}, Vector{Int64}, Vector{Int64}, Vector{Int64}, Vector{Int64}, Vector{Int64}}, Nothing, Nothing, GXBeam.var"#300#302"{Assembly{Float64, Vector{Vector{Float64}}, UnitRange{Int64}, Vector{GXBeam.Element{Float64}}}, Float64, Vector{Int64}, Vector{Int64}, Vector{Int64}, Vector{Int64}, Vector{Int64}, Vector{Int64}}, Nothing, Nothing, Nothing, SparseArrays.SparseMatrixCSC{Bool, Int64}, Nothing, Nothing, Nothing, Nothing, typeof(SciMLBase.DEFAULT_OBSERVED), Nothing}, Vector{Vector{Float64}}, Vector{Float64}, Vector{Vector{Vector{Float64}}}, OrdinaryDiffEq.DABDF2Cache{Vector{Float64}, Vector{Float64}, Vector{Float64}, OrdinaryDiffEq.NLSolver{NLNewton{Rational{Int64}, Rational{Int64}, Rational{Int64}}, true, Vector{Float64}, Rational{Int64}, Nothing, Float64, OrdinaryDiffEq.NLNewtonCache{Vector{Float64}, Float64, Float64, Vector{Float64}, Matrix{Float64}, Matrix{Float64}, Nothing, Nothing, LinearSolve.LinearCache{Matrix{Float64}, Vector{Float64}, Vector{Float64}, SciMLBase.NullParameters, Nothing, LU{Float64, Matrix{Float64}}, LinearSolve.InvPreconditioner{Diagonal{Float64, Vector{Float64}}}, Diagonal{Float64, Vector{Float64}}, Float64}}}, Float64}}, DiffEqBase.DEStats}, DAEFunction{true, GXBeam.var"#299#301"{Assembly{Float64, Vector{Vector{Float64}}, UnitRange{Int64}, Vector{GXBeam.Element{Float64}}}, Float64, Vector{Int64}, Vector{Int64}, Vector{Int64}, Vector{Int64}, Vector{Int64}, Vector{Int64}}, Nothing, Nothing, GXBeam.var"#300#302"{Assembly{Float64, Vector{Vector{Float64}}, UnitRange{Int64}, Vector{GXBeam.Element{Float64}}}, Float64, Vector{Int64}, Vector{Int64}, Vector{Int64}, Vector{Int64}, Vector{Int64}, Vector{Int64}}, Nothing, Nothing, Nothing, SparseArrays.SparseMatrixCSC{Bool, Int64}, Nothing, Nothing, Nothing, Nothing, typeof(SciMLBase.DEFAULT_OBSERVED), Nothing}, OrdinaryDiffEq.DABDF2Cache{Vector{Float64}, Vector{Float64}, Vector{Float64}, OrdinaryDiffEq.NLSolver{NLNewton{Rational{Int64}, Rational{Int64}, Rational{Int64}}, true, Vector{Float64}, Rational{Int64}, Nothing, Float64, OrdinaryDiffEq.NLNewtonCache{Vector{Float64}, Float64, Float64, Vector{Float64}, Matrix{Float64}, Matrix{Float64}, Nothing, Nothing, LinearSolve.LinearCache{Matrix{Float64}, Vector{Float64}, Vector{Float64}, SciMLBase.NullParameters, Nothing, LU{Float64, Matrix{Float64}}, LinearSolve.InvPreconditioner{Diagonal{Float64, Vector{Float64}}}, Diagonal{Float64, Vector{Float64}}, Float64}}}, Float64}, OrdinaryDiffEq.DEOptions{Float64, Float64, Float64, Float64, PIController{Rational{Int64}}, typeof(DiffEqBase.ODE_DEFAULT_NORM), typeof(opnorm), Nothing, CallbackSet{Tuple{}, Tuple{}}, typeof(DiffEqBase.ODE_DEFAULT_ISOUTOFDOMAIN), typeof(DiffEqBase.ODE_DEFAULT_PROG_MESSAGE), typeof(DiffEqBase.ODE_DEFAULT_UNSTABLE_CHECK), DataStructures.BinaryHeap{Float64, DataStructures.FasterForward}, DataStructures.BinaryHeap{Float64, DataStructures.FasterForward}, Nothing, Nothing, Int64, Tuple{}, Tuple{}, Tuple{}}, Vector{Float64}, Float64, Nothing, OrdinaryDiffEq.DefaultInit}, cache::OrdinaryDiffEq.DImplicitEulerCache{Vector{Float64}, Vector{Float64}, Vector{Float64}, OrdinaryDiffEq.NLSolver{NLNewton{Rational{Int64}, Rational{Int64}, Rational{Int64}}, true, Vector{Float64}, Rational{Int64}, Nothing, Float64, OrdinaryDiffEq.NLNewtonCache{Vector{Float64}, Float64, Float64, Vector{Float64}, Matrix{Float64}, Matrix{Float64}, Nothing, Nothing, LinearSolve.LinearCache{Matrix{Float64}, Vector{Float64}, Vector{Float64}, SciMLBase.NullParameters, Nothing, LU{Float64, Matrix{Float64}}, LinearSolve.InvPreconditioner{Diagonal{Float64, Vector{Float64}}}, Diagonal{Float64, Vector{Float64}}, Float64}}}}, repeat_step::Bool)
      @ OrdinaryDiffEq ~/.julia/packages/OrdinaryDiffEq/iOlmD/src/nlsolve/nlsolve.jl:21
    [5] perform_step!(integrator::OrdinaryDiffEq.ODEIntegrator{DABDF2{12, true, Nothing, NLNewton{Rational{Int64}, Rational{Int64}, Rational{Int64}}, typeof(OrdinaryDiffEq.DEFAULT_PRECS), Val{:forward}, true, nothing}, true, Vector{Float64}, Vector{Float64}, Float64, Tuple{var"#103#105"{Int64}, Dict{Int64, DistributedLoads{Float64}}, Dict{Int64, Vector{PointMass{Float64}}}, StaticArrays.SVector{3, Float64}, StaticArrays.SVector{3, Float64}, StaticArrays.SVector{3, Float64}, StaticArrays.SVector{3, Float64}, StaticArrays.SVector{3, Float64}, StaticArrays.SVector{3, Float64}}, Float64, Float64, Float64, Float64, Vector{Vector{Float64}}, DAESolution{Float64, 2, Vector{Vector{Float64}}, Nothing, Nothing, Nothing, Vector{Float64}, DAEProblem{Vector{Float64}, Vector{Float64}, Tuple{Float64, Float64}, true, Tuple{var"#103#105"{Int64}, Dict{Int64, DistributedLoads{Float64}}, Dict{Int64, Vector{PointMass{Float64}}}, StaticArrays.SVector{3, Float64}, StaticArrays.SVector{3, Float64}, StaticArrays.SVector{3, Float64}, StaticArrays.SVector{3, Float64}, StaticArrays.SVector{3, Float64}, StaticArrays.SVector{3, Float64}}, DAEFunction{true, GXBeam.var"#299#301"{Assembly{Float64, Vector{Vector{Float64}}, UnitRange{Int64}, Vector{GXBeam.Element{Float64}}}, Float64, Vector{Int64}, Vector{Int64}, Vector{Int64}, Vector{Int64}, Vector{Int64}, Vector{Int64}}, Nothing, Nothing, GXBeam.var"#300#302"{Assembly{Float64, Vector{Vector{Float64}}, UnitRange{Int64}, Vector{GXBeam.Element{Float64}}}, Float64, Vector{Int64}, Vector{Int64}, Vector{Int64}, Vector{Int64}, Vector{Int64}, Vector{Int64}}, Nothing, Nothing, Nothing, SparseArrays.SparseMatrixCSC{Bool, Int64}, Nothing, Nothing, Nothing, Nothing, typeof(SciMLBase.DEFAULT_OBSERVED), Nothing}, Base.Pairs{Symbol, Union{}, Tuple{}, NamedTuple{(), Tuple{}}}, Vector{Bool}}, DABDF2{12, true, Nothing, NLNewton{Rational{Int64}, Rational{Int64}, Rational{Int64}}, typeof(OrdinaryDiffEq.DEFAULT_PRECS), Val{:forward}, true, nothing}, OrdinaryDiffEq.InterpolationData{DAEFunction{true, GXBeam.var"#299#301"{Assembly{Float64, Vector{Vector{Float64}}, UnitRange{Int64}, Vector{GXBeam.Element{Float64}}}, Float64, Vector{Int64}, Vector{Int64}, Vector{Int64}, Vector{Int64}, Vector{Int64}, Vector{Int64}}, Nothing, Nothing, GXBeam.var"#300#302"{Assembly{Float64, Vector{Vector{Float64}}, UnitRange{Int64}, Vector{GXBeam.Element{Float64}}}, Float64, Vector{Int64}, Vector{Int64}, Vector{Int64}, Vector{Int64}, Vector{Int64}, Vector{Int64}}, Nothing, Nothing, Nothing, SparseArrays.SparseMatrixCSC{Bool, Int64}, Nothing, Nothing, Nothing, Nothing, typeof(SciMLBase.DEFAULT_OBSERVED), Nothing}, Vector{Vector{Float64}}, Vector{Float64}, Vector{Vector{Vector{Float64}}}, OrdinaryDiffEq.DABDF2Cache{Vector{Float64}, Vector{Float64}, Vector{Float64}, OrdinaryDiffEq.NLSolver{NLNewton{Rational{Int64}, Rational{Int64}, Rational{Int64}}, true, Vector{Float64}, Rational{Int64}, Nothing, Float64, OrdinaryDiffEq.NLNewtonCache{Vector{Float64}, Float64, Float64, Vector{Float64}, Matrix{Float64}, Matrix{Float64}, Nothing, Nothing, LinearSolve.LinearCache{Matrix{Float64}, Vector{Float64}, Vector{Float64}, SciMLBase.NullParameters, Nothing, LU{Float64, Matrix{Float64}}, LinearSolve.InvPreconditioner{Diagonal{Float64, Vector{Float64}}}, Diagonal{Float64, Vector{Float64}}, Float64}}}, Float64}}, DiffEqBase.DEStats}, DAEFunction{true, GXBeam.var"#299#301"{Assembly{Float64, Vector{Vector{Float64}}, UnitRange{Int64}, Vector{GXBeam.Element{Float64}}}, Float64, Vector{Int64}, Vector{Int64}, Vector{Int64}, Vector{Int64}, Vector{Int64}, Vector{Int64}}, Nothing, Nothing, GXBeam.var"#300#302"{Assembly{Float64, Vector{Vector{Float64}}, UnitRange{Int64}, Vector{GXBeam.Element{Float64}}}, Float64, Vector{Int64}, Vector{Int64}, Vector{Int64}, Vector{Int64}, Vector{Int64}, Vector{Int64}}, Nothing, Nothing, Nothing, SparseArrays.SparseMatrixCSC{Bool, Int64}, Nothing, Nothing, Nothing, Nothing, typeof(SciMLBase.DEFAULT_OBSERVED), Nothing}, OrdinaryDiffEq.DABDF2Cache{Vector{Float64}, Vector{Float64}, Vector{Float64}, OrdinaryDiffEq.NLSolver{NLNewton{Rational{Int64}, Rational{Int64}, Rational{Int64}}, true, Vector{Float64}, Rational{Int64}, Nothing, Float64, OrdinaryDiffEq.NLNewtonCache{Vector{Float64}, Float64, Float64, Vector{Float64}, Matrix{Float64}, Matrix{Float64}, Nothing, Nothing, LinearSolve.LinearCache{Matrix{Float64}, Vector{Float64}, Vector{Float64}, SciMLBase.NullParameters, Nothing, LU{Float64, Matrix{Float64}}, LinearSolve.InvPreconditioner{Diagonal{Float64, Vector{Float64}}}, Diagonal{Float64, Vector{Float64}}, Float64}}}, Float64}, OrdinaryDiffEq.DEOptions{Float64, Float64, Float64, Float64, PIController{Rational{Int64}}, typeof(DiffEqBase.ODE_DEFAULT_NORM), typeof(opnorm), Nothing, CallbackSet{Tuple{}, Tuple{}}, typeof(DiffEqBase.ODE_DEFAULT_ISOUTOFDOMAIN), typeof(DiffEqBase.ODE_DEFAULT_PROG_MESSAGE), typeof(DiffEqBase.ODE_DEFAULT_UNSTABLE_CHECK), DataStructures.BinaryHeap{Float64, DataStructures.FasterForward}, DataStructures.BinaryHeap{Float64, DataStructures.FasterForward}, Nothing, Nothing, Int64, Tuple{}, Tuple{}, Tuple{}}, Vector{Float64}, Float64, Nothing, OrdinaryDiffEq.DefaultInit}, cache::OrdinaryDiffEq.DImplicitEulerCache{Vector{Float64}, Vector{Float64}, Vector{Float64}, OrdinaryDiffEq.NLSolver{NLNewton{Rational{Int64}, Rational{Int64}, Rational{Int64}}, true, Vector{Float64}, Rational{Int64}, Nothing, Float64, OrdinaryDiffEq.NLNewtonCache{Vector{Float64}, Float64, Float64, Vector{Float64}, Matrix{Float64}, Matrix{Float64}, Nothing, Nothing, LinearSolve.LinearCache{Matrix{Float64}, Vector{Float64}, Vector{Float64}, SciMLBase.NullParameters, Nothing, LU{Float64, Matrix{Float64}}, LinearSolve.InvPreconditioner{Diagonal{Float64, Vector{Float64}}}, Diagonal{Float64, Vector{Float64}}, Float64}}}}, repeat_step::Bool)
      @ OrdinaryDiffEq ~/.julia/packages/OrdinaryDiffEq/iOlmD/src/perform_step/dae_perform_step.jl:65
[…]
