### 2022-04-06 byuflowlab/GXBeam.jl#77 @Kevin-Mattheus-Moerman
Similar errors (`type Nothing has no field α`) occur for `diffeq.ipynb`
