### 2022-04-06 #3997 @Kevin-Mattheus-Moerman
@taylormcd

- [ ] I get some errors during testing, see; https://github.com/byuflowlab/GXBeam.jl/issues/77
