### 2022-04-06 byuflowlab/GXBeam.jl#77 @taylormcd
This is a duplicate of #68 which was caused by regressions in DifferentialEquations.  See https://github.com/SciML/DifferentialEquations.jl/issues/846.  Are you using v0.3.2?
