+++ Project.toml
@@ -22,7 +22,7 @@ ForwardDiff = "0.10"
 LinearMaps = "2, 3"
 NLsolve = "4"
 OffsetArrays = "1"
-Roots = "1"
+Roots = "1, 2"
 SciMLBase = "1"
 StaticArrays = "0.12, 1.0"
 WriteVTK = "1"
