### 2022-04-11 534f7f9
docstrings.  some refactoring
