+++ src/section.jl
@@ -1,6 +1,22 @@
 using LinearAlgebra: I, Symmetric, det, factorize
 using SparseArrays: spzeros, sparse
 
+"""
+    Material(E1, E2, E2, nu12, nu13, nu23, G12, G13, G23, rho)
+
+General orthotropic material properties. 
+The "1" direction is along the beam axis for a fiber orientation of zero (theta=0).
+"2" corresponds to the local x-direction and "3" the local y at theta=0. (see documentation for figures).  
+If the two in-plane directions have the same stiffness properties then only one 
+needs to be specified (e.g., compatible with the plane-stress assumption of CLT-based tools):
+`Material(E1, E2, nu12, G12, rho)`
+
+**Arguments**
+- `E::float`: Young's modulus along 1st, 2nd and 3rd axes.
+- `nu::float`: Poisson's ratio.  nu_ij E_j = nu_ji E_i
+- `G::float`: shear moduli
+- `rho::float`: density
+"""
 struct Material{TF}
     E1::TF
     E2::TF
@@ -16,18 +32,42 @@ end
 
 Material(E1, E2, nu12, G12, rho) = Material(E1, E2, E2, nu12, nu12, nu12, G12, G12, G12, rho)
 
+"""
+    Node(x, y, number)
+
+A node in the finite element mesh at location x, y with a given index number.
+
+**Arguments**
+- `x::float`: x location of node in global coordinate system
+- `y::float`: y location of node in global coordinate system
+- `number::integer`: unique index of this node, elements use this indices
+"""
 struct Node{TF, TI}
     x::TF
     y::TF
     number::TI
 end
 
+"""
+    Element(nodenum, material, theta)
+
+An element in the mesh, consisting of four ordered nodes, a material, and a fiber orientation.
+
+**Arguments**
+- `nodenum::Vector{integer}`: a vector of four node numbers corresponding the the four nodes defining this element. 
+    Node order should be counterclockwise starting from the bottom left node using the local coordinate sytem (see figure).
+- `material::Material`: material properties of this element
+- `theta::float`: fiber orientation
+"""
 struct Element{VI, TF}
     nodenum::VI
     material::Material
     theta::TF
 end
 
+"""
+Constituitive matrix of this material using the internal ordering.
+"""
 function stiffness(material) 
     E1 = material.E1; E2 = material.E2; E3 = material.E3
     nu12 = material.nu12; nu13 = material.nu13; nu23 = material.nu23
@@ -52,6 +92,9 @@ function stiffness(material)
     return Symmetric(Q)
 end
 
+"""
+Rotate constituitive matrix by ply angle
+"""
 function rotate_to_ply(K, theta)
     c = cos(theta)
     s = sin(theta)
@@ -65,6 +108,9 @@ function rotate_to_ply(K, theta)
     return T*K*T'
 end
 
+"""
+Rotate constituitive matrix by element orientation where `c = cos(beta)` and `s = sin(beta)`
+"""
[…]
