+++ src/analyses.jl
@@ -35,47 +35,22 @@ iteration procedure converged.
  - `tvec = 0`: Time vector/value. May be used in conjunction with time varying
         prescribed conditions and distributed loads to gradually increase
         displacements/loads.
+ - `show_trace = false`: Flag indicating whether to show solution progress
  - `reset_state = true`: Flag indicating whether the state variables should be
         reset prior to performing the analysis.  This keyword argument is only valid
         for the pre-allocated version of this function.
 """
-function static_analysis(assembly;
-    prescribed_conditions=Dict{Int,PrescribedConditions{Float64}}(),
-    distributed_loads=Dict{Int,DistributedLoads{Float64}}(),
-    point_masses=Dict{Int,PointMass{Float64}}(),
-    gravity=SVector(0,0,0),
-    linear=false,
-    linearization_state=nothing,
-    update_linearization_state=false,
-    method=:newton,
-    linesearch=LineSearches.BackTracking(maxstep=1e6),
-    ftol=1e-9,
-    iterations=1000,
-    tvec=0.0,
-    )
+function static_analysis(assembly; kwargs...)
 
     system = System(assembly)
 
-    return static_analysis!(system, assembly;
-        prescribed_conditions=prescribed_conditions,
-        distributed_loads=distributed_loads,
-        point_masses=point_masses,
-        gravity=gravity,
-        linear=linear,
-        linearization_state=linearization_state,
-        update_linearization_state=update_linearization_state,
-        method=method,
-        linesearch=linesearch,
-        ftol=ftol,
-        iterations=iterations,
-        tvec=tvec,
-        reset_state=false)
+    return static_analysis!(system, assembly; kwargs..., reset_state=false)
 end
 
 """
     static_analysis!(system, assembly; kwargs...)
 
-Pre-allocated version of `static_analysis`.
+Pre-allocated version of [`static_analysis`](@ref).
 """
 function static_analysis!(system, assembly;
     prescribed_conditions=Dict{Int,PrescribedConditions{Float64}}(),
@@ -90,6 +65,7 @@ function static_analysis!(system, assembly;
     ftol=1e-9,
     iterations=1000,
     tvec=0.0,
+    show_trace=false,
     reset_state=true)
 
     # reset state, if specified
@@ -110,6 +86,10 @@ function static_analysis!(system, assembly;
     
     # begin time stepping
     for t in tvec
+
+        if show_trace
+            println("Solving for t=$t")
+        end
 
         # update stored time
         system.t = t
@@ -149,9 +129,7 @@ function static_analysis!(system, assembly;
             df = NLsolve.OnceDifferentiable(f!, j!, xs, rs, Ks)
 
             result = NLsolve.nlsolve(df, xs,
-            # result = NLsolve.nlsolve(f!, xs,
-                # autodiff=:forward,
-                show_trace=true,
+                show_trace=show_trace,
                 linsolve=(x, A, b) -> ldiv!(x, safe_lu(A), b),
                 method=method,
                 linesearch=linesearch,
@@ -203,8 +181,9 @@ iteration procedure converged.
  - `update_linearization_state`: Flag indicating whether to update the linearization state 
     variables for a linear analysis with the current state variables.
  - `method = :newton`: Method (as defined in NLsolve) to solve nonlinear system of equations
- - `linesearch = LineSearches.LineSearches.BackTracking(maxstep=1e6)`: Line search used to solve nonlinear system of equations
- - `ftol = 1e-9`: tolerance for solving nonlinear system of equations
[…]
