+++ test/runtests.jl
@@ -85,29 +85,30 @@ end
     # point masses
     pmass = Dict(
         # point mass at the end of the beam
-        nelem => PointMass(Symmetric(rand(RNG,6,6)))
+        nelem+1 => PointMass(Symmetric(1e3*rand(RNG,6,6)))
     )
 
     # gravity vector
-    gvec = rand(RNG, 3)
+    gvec = 1e3*rand(RNG, 3)
 
     system = System(assembly)
+
     force_scaling = system.force_scaling
+    indices = system.static_indices
+    x = rand(RNG, length(system.x))
+    J = similar(x, length(x), length(x))
+    xs = GXBeam.get_static_state(system, x)
+    Js = similar(xs, length(xs), length(xs))
    
     # --- Static Analysis --- #
-
-    indices = system.static_indices
-
-    x = rand(RNG, 6*length(assembly.elements) + 6*length(assembly.points))
-    J = similar(x, length(x), length(x))
 
     f = (x) -> GXBeam.static_system_residual!(similar(x), x, indices, force_scaling, 
         assembly, pcond, dload, pmass, gvec)
 
-    GXBeam.static_system_jacobian!(J, x, indices, force_scaling,
+    GXBeam.static_system_jacobian!(Js, xs, indices, force_scaling,
         assembly, pcond, dload, pmass, gvec)
 
-    @test all(isapprox.(J, ForwardDiff.jacobian(f, x), atol=1e-10))
+    @test all(isapprox.(Js, ForwardDiff.jacobian(f, xs), atol=1e-10))
 
     # --- Steady State Analysis --- #
 
@@ -118,9 +119,6 @@ end
     ω0 = rand(RNG, 3)
     a0 = rand(RNG, 3)
     α0 = rand(RNG, 3)
-
-    x = rand(RNG, length(system.x))
-    J = similar(x, length(x), length(x))
 
     f = (x) -> GXBeam.steady_state_system_residual!(similar(x), x, indices, force_scaling, 
         assembly, pcond, dload, pmass, gvec, x0, v0, ω0, a0, α0)
@@ -1342,14 +1340,14 @@ end
 
     # combine elements and points into one array
     nelem = nelem_b1 + nelem_b12 + nelem_b2
-    points = vcat(xp_b1, xp_b2[2:end]) # don't duplicate points
+    points = vcat(xp_b1, xp_b2)
     lengths = vcat(lengths_b1, lengths_b12, lengths_b2)
     midpoints = vcat(xm_b1, xm_b12, xm_b2)
     Cab = vcat(Cab_b1, Cab_b12, Cab_b2)
 
     # specify connectivity
-    start = vcat(1:nelem_b1+1, nelem_b1+1:nelem_b1+nelem_b2)
-    stop = vcat(2:nelem_b1+1, nelem_b1+1:nelem_b1+nelem_b2+1)
+    start = 1:nelem
+    stop = 2:nelem+1
 
     # cross section
     w = 1 # inch
@@ -1402,55 +1400,9 @@ end
     @test converged
 end
 
-@testset "Element Gravitational Loads" begin
-
-    # use arbitrary length
-    ΔL = rand(RNG)
-
-    # use random rotation matrix  
-    CtCab = GXBeam.get_C(rand(RNG, 3))
-
-    # create random mass matrix
-    μ = rand(RNG)
-    xm2 = rand(RNG)
-    xm3 = rand(RNG)
-    i22 = rand(RNG)
-    i33 = rand(RNG)
-    i23 = rand(RNG)
-
-    mass = [
-        μ 0 0 0 μ*xm3 -μ*xm2; 
-        0 μ 0 -μ*xm3 0 0; 
-        0 0 μ μ*xm2 0 0; 
[…]
