+++ src/postprocess.jl
@@ -4,20 +4,20 @@
 Holds the state variables for a point
 
 # Fields:
- - `F`: Externally applied forces on the point
- - `M`: Externally applied moments on the point
- - `u`: Displacement variables for the point
- - `theta`: Displacement Wiener-Milenkovic rotation parameters for the point
- - `V`: Linear velocity of the point
- - `Ω`: Angular velocity of the point
+ - `u`: Linear deflection
+ - `theta`: Angular deflection (Wiener-Milenkovic parameters)
+ - `V`: Linear velocity
+ - `Ω`: Angular velocity
+ - `F`: External forces
+ - `M`: External moments
 """
 struct PointState{TF}
     u::SVector{3, TF}
     theta::SVector{3, TF}
-    F::SVector{3, TF}
-    M::SVector{3, TF}
     V::SVector{3, TF}
     Omega::SVector{3, TF}
+    F::SVector{3, TF}
+    M::SVector{3, TF}
 end
 
 """
@@ -26,20 +26,20 @@ end
 Holds the state variables for an element
 
 # Fields:
- - `u`: Displacement variables for the element
- - `theta`: Wiener-Milenkovic rotational displacement variables for the element
- - `F`: Resultant forces for the element (in the deformed local coordinate frame)
- - `M`: Resultant moments for the element (in the deformed local coordinate frame)
- - `V`: Linear velocity of the element (in the deformed local coordinate frame)
- - `Omega`: Angular velocity of the element (in the deformed local coordinate frame)
+ - `u`: Linear deflection
+ - `theta`: Angular deflection
+ - `V`: Linear velocity
+ - `Omega`: Angular velocity
+ - `Fi`: Internal forces
+ - `Mi`: Internal moments
 """
 struct ElementState{TF}
     u::SVector{3, TF}
     theta::SVector{3, TF}
-    F::SVector{3, TF}
-    M::SVector{3, TF}
     V::SVector{3, TF}
     Omega::SVector{3, TF}
+    Fi::SVector{3, TF}
+    Mi::SVector{3, TF}
 end
 
 """
@@ -98,18 +98,18 @@ function extract_point_state(system, assembly, ipoint, x = system.x;
     pc = typeof(prescribed_conditions) <: AbstractDict ? prescribed_conditions : prescribed_conditions(t)
 
     # extract point state variables
-    u, theta = point_displacement(x, ipoint, dynamic_indices.icol_point, pc)
-    F, M = point_loads(x, ipoint, dynamic_indices.icol_point, force_scaling, pc)
+    u, θ = point_displacement(x, ipoint, dynamic_indices.icol_point, pc)
     V, Ω = point_velocities(x, ipoint, dynamic_indices.icol_point)
+    F, M = point_loads(x, ipoint, dynamic_indices.icol_point, force_scaling, pc)
 
     # convert rotation parameter to Wiener-Milenkovic parameters
-    scaling = rotation_parameter_scaling(theta)
-    theta *= scaling
+    scaling = rotation_parameter_scaling(θ)
+    θ *= scaling
 
     # promote all variables to the same type
-    u, theta, F, M, V, Ω = promote(u, theta, F, M, V, Ω)
+    u, θ, V, Ω, F, M = promote(u, θ, V, Ω, F, M)
 
-    return PointState(u, theta, F, M, V, Ω)
+    return PointState(u, θ, V, Ω, F, M)
 end
 
 """
@@ -181,9 +181,9 @@ function extract_element_state(system, assembly, ielem, x = system.x;
     θ *= scaling
 
     # promote all variables to the same type
-    u, θ, F, M, V, Ω = promote(u, θ, F, M, V, Ω)
[…]
