+++ src/index-reduction.md
@@ -1,548 +1,0 @@
-# Index Reduction
-
-To simulate the behavior of a given set of beam elements and nodes, this package constructs and solves a set of differential algebraic equations.  While these equations may be solved as is, it can be sometimes be more convenient or efficient to express a given set of differential algebraic equations as ordinary differential equations.  We would therefore like to be able to represent the differential algebraic equations modeled in this package as ordinary differential equations.  
-
-We first consider the following equations for a single cantilevered beam element.
-
-$$
-f_{u_1}^{-} - F_1^* = 0 \\
-f_{\psi_1}^{-} - M_1^* = 0 \\ 
-f_{V_1} = 0 \\
-f_{\Omega_1} = 0 \\
-\hat{u}_2 - \hat{u}_1 - \Delta u = 0 \\
-\hat{\theta}_2 - \hat{\theta}_1 - \Delta \theta = 0 \\
-f_{u_1}^{+} - F_{N+1}^* = 0 \\
-f_{\psi_1}^{+} - M_{N+1}^* = 0 \\ 
-f_{V_2} = 0 \\
-f_{\Omega_2} = 0 \\
-$$
-where 
-$$
-f_{u_i}^{\mp} = 
-\mp C^T C^{ab} F_i
-- \bar{f}_i^{\mp} 
-+ \frac{\Delta L}{2} 
-[ \tilde{\omega} C^T C^{ab} P_i + \dot{\overline{C^T C^{ab} P_i}}] 
-\\
-f_{\psi_i}^{\mp} = 
-\mp C^T C^{ab} M_i 
-- \bar{m}_i^{\mp} 
-+ \frac{\Delta L}{2} 
-[ 
-  \tilde{\omega} C^T C^{ab} H_i 
-  + \dot{\overline{C^T C^{ab} H_i}}
-  + C^T C^{ab} (\tilde{V}_i P_i - (\tilde{e}_i + \tilde{\gamma}_i) F_i) 
-] \\
-\Delta u = \Delta L \left(C^T C^{ab} (e_i + \gamma_i) - C^{ab} e_i\right) \\
-\Delta \theta = \Delta L \left( Q_a^{-1} C^{ab} \kappa_i \right) \\
-f_{V_i} = C^T C^{ab} V_i - v_i - \tilde{\omega_a} u_i - \dot{u}_i \\
-f_{\Omega_i} = \Omega_i - C^{ba} C \omega_a - C^{ba} Q_a \dot{\theta} \\
-\gamma = S^{11}_1 F_i + S^{12}_1 M_i \\
-\kappa = S^{21}_i F_i + S^{22}_1 M_i \\
-P = M^{11}_i V_i + M^{12}_i \Omega_i \\
-H = M^{21}_1 V_i + M^{22}_1 \Omega_i \\
-$$
-
-The last two governing equations can be manipulated to define explicit expressions for $\dot{u}$ and $\dot{\theta}$.
-
-$$
[…]
