### 2022-04-19 075e214
preallocate matrices and modify in place for big speed up
