+++ src/section.jl
@@ -1,36 +1,37 @@
 using LinearAlgebra: I, Symmetric, det, factorize
 using SparseArrays: spzeros, sparse
+using StaticArrays
 
 """
-    Material(E1, E2, E2, nu12, nu13, nu23, G12, G13, G23, rho)
+    Material(E1, E2, E2, G12, G13, G23, nu12, nu13, nu23, rho)
 
 General orthotropic material properties. 
 The "1" direction is along the beam axis for a fiber orientation of zero (theta=0).
 "2" corresponds to the local x-direction and "3" the local y at theta=0. (see documentation for figures).  
 If the two in-plane directions have the same stiffness properties then only one 
 needs to be specified (e.g., compatible with the plane-stress assumption of CLT-based tools):
-`Material(E1, E2, nu12, G12, rho)`
+`Material(E1, E2, G12, nu12, rho)`
 
 **Arguments**
 - `E::float`: Young's modulus along 1st, 2nd and 3rd axes.
-- `nu::float`: Poisson's ratio.  nu_ij E_j = nu_ji E_i
 - `G::float`: shear moduli
+- `nu::float`: Poisson's ratio.  nu_ij E_j = nu_ji E_i
 - `rho::float`: density
 """
 struct Material{TF}
     E1::TF
     E2::TF
     E3::TF
-    nu12::TF
-    nu13::TF
-    nu23::TF
     G12::TF
     G13::TF
     G23::TF
+    nu12::TF
+    nu13::TF
+    nu23::TF
     rho::TF
 end
 
-Material(E1, E2, nu12, G12, rho) = Material(E1, E2, E2, nu12, nu12, nu12, G12, G12, G12, rho)
+Material(E1, E2, G12, nu12, rho) = Material(E1, E2, E2, G12, G12, G12, nu12, nu12, nu12, rho)
 
 """
     Node(x, y, number)
@@ -66,9 +67,156 @@ struct Element{VI, TF}
 end
 
 """
+internal cache so allocations happen only once upfront
+"""
+struct Cache{TM, TSM, TFM, TV}
+    Q::TM
+    Ttheta::TSM
+    Tbeta::TSM
+    Z::TSM
+    S::TFM
+    N::TSM
+    SZ::TSM
+    SN::TSM
+    Bksi::TSM
+    Beta::TSM
+    dNM_dksi::TSM
+    dNM_deta::TSM
+    BN::TSM
+    Ae::TM
+    Re::TM
+    Ee::TM
+    Ce::TM
+    Le::TM
+    Me::TM
+    idx::TV
+end
+
+"""
+create chace.  set sizes of static matrices, and set sparsity patterns for those that are fixed.
+"""
+function initializecache(etype=Float64)
+
+    # create cache (TODO: set sparsity patterns)
+    Q = zeros(etype, 6, 6)
+    
+    Ttheta = zeros(etype, 6, 6)
+    Ttheta[1, 1] = 1.0; Ttheta[1, 4] = 1.0; Ttheta[1, 6] = 1.0
+    Ttheta[2, 2] = 1.0
+    Ttheta[3, 3] = 1.0; Ttheta[3, 5] = 1.0
+    Ttheta[4, 1] = 1.0; Ttheta[4, 4] = 1.0; Ttheta[4, 6] = 1.0
[…]
