### 2022-04-19 166cdd6
seprate out plotting function
