### 2022-04-20 74bf615
add theory for structural damping
