+++ docs/src/theory.md
@@ -1,0 +1,78 @@
+# Structural Damping
+
+When stiffness proportional structural damping is used, the relationship between stresses and strains takes the following form:
+
+```math
+\begin{bmatrix} F \\  M \end{bmatrix} = S \begin{bmatrix} \gamma \\ \kappa \end{bmatrix} + \mu S \begin{bmatrix} \dot{\gamma} \\ \dot{\kappa} \end{bmatrix}  \\
+```
+
+where ``\mu`` is a diagonal matrix with damping coefficients.
+
+Rearranging this equations yields a new expression for the strains and curvatures when stiffness proportional structural damping is assumed.
+
+```math
+\begin{bmatrix} \gamma \\ \kappa \end{bmatrix} = S^{-1} \begin{bmatrix} F \\  M \end{bmatrix} - \mu \begin{bmatrix} \dot{\gamma} \\ \dot{\kappa} \end{bmatrix}
+```
+
+## Strain Rates
+
+The strain rates may be expressed as a function of our state variables using the strain compatability relationship.  We first rearrange the strain compatability equation to yield the following expression for the element strain.
+
+```math
+\gamma_i = C^{ba} C \frac{\Delta u}{\Delta L} + C^{ba} C C^{ab} e_1 - e_1 
+```
+
+We then differentiate this equation with respect to time to obtain an expression for the strain rates:
+
+```math
+\dot{\gamma}_i = C^{ba} C \frac{\Delta \dot{u}}{\Delta L} + C^{ba} \dot{C} \frac{\Delta u}{\Delta L} + C^{ba} \dot{C} C^{ab} e_1 
+```
+
+We can then use the identity ``C^{ba} \dot{C} = -\widetilde{(\Omega-\omega)} C^{ba} C`` to express the strain rates as a function of our state variables. 
+
+```math
+\dot{\gamma}_i = C^{ba} C \frac{\Delta \dot{u}}{\Delta L} - \widetilde{(\Omega-\omega)} C^{ba} C \frac{\Delta u}{\Delta L} -  \widetilde{(\Omega-\omega)} C^{ba} C C^{ab} e_1 
+```
+
+## Curvature Rates
+
+The curvature rates may be expressed as a function of our state variables using the curvature compatability relationship.  We first rearrange the curvature compatability equation to yield the following expression for the element curvature.
+
+```math
+\kappa_i = C^{ba} Q \frac{\Delta \theta}{\Delta L}
+```
+
+Differentiating this equation with respect to time yields:
+
+```math
+\dot{\kappa}_i = C^{ba} Q \frac{\Delta \dot{\theta}}{\Delta L} + C^{ba} \dot{Q} \frac{\Delta \theta}{\Delta L}
+```
+
+We can expand ``\dot{Q}`` using its partial derivatives 
+
+```math
+\dot{Q} = Q_{\theta_1} \dot{\theta_1} + Q_{\theta_2} \dot{\theta_2} + Q_{\theta_3} \dot{\theta_3}
+```
+
+so that
+
+```math
+\dot{Q} \Delta \theta = 
+(Q_{\theta_1} \Delta \theta) \dot{\theta_1} + 
+(Q_{\theta_2} \Delta \theta) \dot{\theta_2} + 
+(Q_{\theta_3} \Delta \theta) \dot{\theta_3} 
+\equiv \Delta Q \dot{\theta}
+```
+
+Our expression for the curvature rates is then
+
+```math
+\dot{\kappa}_i = C^{ba} Q \frac{\Delta \dot{\theta}}{\Delta L} + C^{ba} \Delta Q \frac{\dot{\theta}}{\Delta L}
+```
+
[…]
