### 2022-04-21 af3c968
finish up structural damping
