+++ docs/src/literate/wind-turbine-blade.jl
@@ -58,7 +58,8 @@ end
 t = 0:0.001:2.0
 
 system, history, converged = time_domain_analysis(assembly, t;
-    prescribed_conditions = prescribed_conditions)
+    prescribed_conditions = prescribed_conditions,
+    structural_damping = false)
 
 #!jl nothing #hide
 
