+++ src/element.jl
@@ -110,19 +110,34 @@ analysis
 end
 
 """
-    steady_state_element_properties(x, indices, force_scaling, assembly, ielem, 
-        prescribed_conditions, gravity, x0, v0, ω0, a0, α0)
+    steady_state_element_properties(x, indices, force_scaling, structural_damping, 
+        assembly, ielem, prescribed_conditions, gravity, x0, v0, ω0, a0, α0)
 
 Calculate/extract the element properties needed to construct the residual for a steady 
 state analysis
 """
-@inline function steady_state_element_properties(x, indices, force_scaling, assembly, ielem, 
-    prescribed_conditions, gravity, x0, v0, ω0, a0, α0)
+@inline function steady_state_element_properties(x, indices, force_scaling, structural_damping, 
+    assembly, ielem, prescribed_conditions, gravity, x0, v0, ω0, a0, α0)
 
     properties = static_element_properties(x, indices, force_scaling, assembly, ielem, 
         prescribed_conditions, gravity)
 
-    @unpack CtCab, mass11, mass12, mass21, mass22, u, θ, a, α = properties
+    @unpack L, Cab, C, CtCab, Qinv, mass11, mass12, mass21, mass22, u1, u2, θ1, θ2, 
+        u, θ, γ, κ, a, α = properties
+
+    # damping coefficients
+    μ = assembly.elements[ielem].mu
+
+    # damping submatrices
+    μ11 = @SMatrix [μ[1] 0 0; 0 μ[2] 0; 0 0 μ[3]]
+    μ22 = @SMatrix [μ[4] 0 0; 0 μ[5] 0; 0 0 μ[6]]
+
+    # rotation parameter matrices
+    C1 = get_C(θ1)
+    C2 = get_C(θ2)
+    Qinv1 = get_Qinv(θ1)
+    Qinv2 = get_Qinv(θ2)
+    Q = get_Q(θ)
 
     # linear and angular velocity
     V1, Ω1 = point_velocities(x, assembly.start[ielem], indices.icol_point)
@@ -135,31 +150,74 @@ state analysis
     H = CtCab*mass21*CtCab'*V + CtCab*mass22*CtCab'*Ω
 
     # distance from the rotation center
+    Δx1 = assembly.points[assembly.start[ielem]] - x0
+    Δx2 = assembly.points[assembly.stop[ielem]] - x0
     Δx = assembly.elements[ielem].x - x0
 
-    # undeformed linear and angular velocity
+    # undeformed linear velocity
+    v1 = v0 + cross(ω0, Δx1 - x0)
+    v2 = v0 + cross(ω0, Δx2 - x0)
     v = v0 + cross(ω0, Δx)
+
+    # undeformed angular velocity
+    ω1 = ω0
+    ω2 = ω0
     ω = ω0
 
     # linear and angular acceleration
     a += a0 + cross(α0, Δx) + cross(α0, u)
     α += α0
 
-    return (; properties..., V1, V2, Ω1, Ω2, V, Ω, P, H, v, ω, a, α) 
+    # linear deflection rates 
+    udot1 = V1 - v1 - cross(ω1, u1)
+    udot2 = V2 - v2 - cross(ω2, u2)
+    udot = (udot1 + udot2)/2
+
+    # angular deflection rates
+    θdot1 = Qinv1*C1*(Ω1 - ω1)
+    θdot2 = Qinv2*C2*(Ω2 - ω2)
+    θdot = (θdot1 + θdot2)/2
+
+    # change in linear and angular displacement
+    Δu = u2 - u1
+    Δθ = θ2 - θ1
+
+    # change in linear and angular displacement rates
[…]
