+++ src/GXBeam.jl
@@ -50,7 +50,7 @@ include("system.jl")
 include("analyses.jl")
 include("postprocess.jl")
 
-include("interfaces/differentialequations.jl")
+include("interfaces/diffeq.jl")
 include("interfaces/forwarddiff.jl")
 
 end
