+++ docs/src/literate/overdetermined.jl
@@ -157,7 +157,7 @@ plot(
 
 x = [assembly.elements[ielem].x[1] + state.elements[ielem].u[1] for ielem =
     1:length(assembly.elements)]
-M = [state.elements[ielem].M[2] for ielem = 1:length(assembly.elements)]
+M = [state.elements[ielem].Mi[2] for ielem = 1:length(assembly.elements)]
 plot!(x_a, M_a, label="Analytical")
 scatter!(x, M, label="GXBeam")
 #!nb plot!(show=true)
