### 2022-04-22 #3997 @taylormcd
I added a short contributing section to README.md.  I'll probably create a more detailed one later, but this should suffice for now.
