+++ src/point.jl
@@ -153,6 +153,161 @@ end
 end
 
 """
+    expanded_point_loads(x, ipoint, icol, force_scaling, prescribed_conditions)
+
+Extract the loads `CF` and `CM` of point `ipoint` from the state variable vector or 
+prescribed conditions for a constant mass matrix system.
+"""
+@inline function expanded_point_loads(x, ipoint, icol, force_scaling, prescribed_conditions)
+
+    if haskey(prescribed_conditions, ipoint)
+        F, M = expanded_point_loads(x, icol[ipoint], force_scaling, prescribed_conditions[ipoint])
+    else
+        F, M = expanded_point_loads(x, icol[ipoint], force_scaling)
+    end
+
+    return F, M
+end
+
+@inline function expanded_point_loads(x, icol, force_scaling, prescribed_conditions)
+
+    # unpack prescribed conditions for the node
+    @unpack value, follower, isforce = prescribed_conditions
+
+    # displacements
+    u, θ = point_displacement(x, icol, prescribed_conditions)
+
+    # transformation matrix
+    C = get_C(θ)
+
+    # calculate prescribed external forces
+    CF = zero(u)
+    for i = 1:3
+        if isforce[i]
+            # add dead force
+            CF += SVector(C[1,i], C[2,i], C[3,i])*value[i]
+            # add follower force
+            CF += SVector(I3[1,i], I3[2,i], I3[3,i])*follower[i]
+        end
+    end
+
+    # overwrite prescribed external forces if linear displacements are prescribed instead
+    CF = SVector(ifelse(isforce[1], CF[1], x[icol  ]*force_scaling),
+                 ifelse(isforce[2], CF[2], x[icol+1]*force_scaling),
+                 ifelse(isforce[3], CF[3], x[icol+2]*force_scaling))
+
+    # calculate prescribed external moments
+    CM = zero(θ)
+    for i = 4:6
+        if isforce[i]
+            # add dead moment
+            CM += SVector(C[1,i-3], C[2,i-3], C[3,i-3])*value[i]
+            # add follower moment
+            CM += SVector(I3[1,i-3], I3[2,i-3], I3[3,i-3])*follower[i]
+        end
+    end
+
+    # overwrite prescribed external moments if angular displacements are prescribed instead
+    CM = SVector(ifelse(isforce[4], CM[1], x[icol+3]*force_scaling),
+                 ifelse(isforce[5], CM[2], x[icol+4]*force_scaling),
+                 ifelse(isforce[6], CM[3], x[icol+5]*force_scaling))
+
+    return CF, CM
+end
+
+@inline function expanded_point_loads(x, icol, force_scaling)
+
+    CF = @SVector zeros(eltype(x), 3)
+    CM = @SVector zeros(eltype(x), 3)
+
+    return CF, CM
+end
+
+"""
+    expanded_point_load_jacobians(x, ipoint, icol, force_scaling, prescribed_conditions)
+
+Calculate the load jacobians `CF_θ`, `CF_CF`, `CM_θ`, and `CM_CM` of point `ipoint` for a 
+constant mass matrix system.
+"""
+@inline function expanded_point_load_jacobians(x, ipoint, icol, force_scaling, prescribed_conditions)
+
[…]
