+++ src/interfaces/expanded.jl
@@ -1,199 +1,0 @@
-"""
-    expanded_element_loads(x, ielem, icol_elem, force_scaling)
-
-Extract the internal loads of a beam element at its endpoints (`F1`, `M1`, `F2`, `M2`) 
-from the state variable vector for the expanded system.  These loads are expressed in the 
-deformed element frame.
-"""
-@inline function expanded_element_loads(x, ielem, icol_elem, force_scaling)
-
-    icol = icol_elem[ielem]
-
-    F1 = SVector(x[icol  ], x[icol+1], x[icol+2]) .* force_scaling
-    M1 = SVector(x[icol+3], x[icol+4], x[icol+5]) .* force_scaling
-    F2 = SVector(x[icol+6], x[icol+7], x[icol+8]) .* force_scaling
-    M2 = SVector(x[icol+9], x[icol+10], x[icol+11]) .* force_scaling
-
-    return F1, M1, F2, M2
-end
-
-"""
-    expanded_element_velocities(x, ielem, icol_elem, force_scaling)
-
-Extract the velocities of a beam element from the state variable vector of the expanded
-system.  These velocities are expressed in the deformed frame of reference.
-"""
-@inline function expanded_element_velocities(x, ielem, icol_elem)
-
-    icol = icol_elem[ielem]
-
-    V = SVector(x[icol+12], x[icol+13], x[icol+14])
-    Ω = SVector(x[icol+15], x[icol+16], x[icol+17])
-
-    return V, Ω
-end
-
-"""
-    expanded_velocity_residuals(properties)
-
-Calculate the velocity residuals `rV` and `rΩ` for a beam element
-"""
-@inline function expanded_velocity_residuals(properties)
-
-    @unpack C, C1, C2, V, V1, V2, Ω, Ω1, Ω2 = properties
-    
-    rV = C'*V - (C1'*V1 + C2'*V2)/2
-    rΩ = C'*Ω - (C1'*Ω1 + C2'*Ω2)/2
-
-    return (; rV, rΩ)
-end
-
-"""
-    expanded_equilibrium_residuals(properties, distributed_loads, ielem)
-
-Calculate the equilibrium residuals for a beam element for a static analysis.
-"""
-@inline function static_equilibrium_residuals(properties, distributed_loads, ielem)
-  
-    @unpack L, Cab, mass11, mass12, mass21, mass22, C, CtCab, F1, F2, M1, M2, F, M, γ, κ, a, α = properties
-
-    P = mass11*V + mass12*Ω
-    H = mass21*V + mass22*Ω
-
-    Pdot = mass11*Vdot + mass12*Ωdot
-    Hdot = mass21*Vdot + mass22*Ωdot
-
-    # initialize equilibrium residual
-    rF = Cab*F2 - Cab*F1
-    rM = Cab*M2 - Cab*M1
-
-    # add loads due to internal forces/moments and stiffness
-    rM += Cab*cross(L*e1 + γ, F)
-
-    # add loads due to linear and angular acceleration (including gravity)
-    rF -= mass11*C*a + mass12*C*α
-    rM -= mass21*C*a + mass22*C*α
-
[…]
