+++ src/element.jl
@@ -55,6 +55,40 @@ These loads are expressed in the deformed element frame.
 end
 
 """
+    expanded_element_loads(x, ielem, icol_elem, force_scaling)
+
+Extract the internal loads of a beam element at its endpoints (`F1`, `M1`, `F2`, `M2`) 
+from the state variable vector for a constant mass matrix system.
+"""
+@inline function expanded_element_loads(x, ielem, icol_elem, force_scaling)
+
+    icol = icol_elem[ielem]
+
+    F1 = SVector(x[icol  ], x[icol+1], x[icol+2]) .* force_scaling
+    M1 = SVector(x[icol+3], x[icol+4], x[icol+5]) .* force_scaling
+    F2 = SVector(x[icol+6], x[icol+7], x[icol+8]) .* force_scaling
+    M2 = SVector(x[icol+9], x[icol+10], x[icol+11]) .* force_scaling
+
+    return F1, M1, F2, M2
+end
+
+"""
+    expanded_element_velocities(x, ielem, icol_elem, force_scaling)
+
+Extract the velocities of a beam element from the state variable vector for a constant mass
+matrix system.
+"""
+@inline function expanded_element_velocities(x, ielem, icol_elem)
+
+    icol = icol_elem[ielem]
+
+    CV = SVector(x[icol+12], x[icol+13], x[icol+14])
+    CΩ = SVector(x[icol+15], x[icol+16], x[icol+17])
+
+    return CV, CΩ
+end
+
+"""
     static_element_properties(x, indices, force_scaling, assembly, ielem, 
         prescribed_conditions, gravity)
 
@@ -414,6 +448,128 @@ analysis
 end
 
 """
+    expanded_element_properties(x, indices, force_scaling, structural_damping, 
+        assembly, ielem, prescribed_conditions, gravity, x0, v0, ω0, a0, α0)
+
+Calculate/extract the element properties needed to construct the residual for a constant
+mass matrix system
+"""
+@inline function expanded_element_properties(x, indices, force_scaling, structural_damping, 
+    assembly, ielem, prescribed_conditions, gravity, x0, v0, ω0, a0, α0)
+
+    # unpack element parameters
+    @unpack L, Cab, compliance, mass, mu = assembly.elements[ielem]
+
+    # scale compliance and mass matrices by the element length
+    compliance *= L
+    mass *= L
+
+    # compliance submatrices (in the deformed element frame)
+    S11 = compliance[SVector{3}(1:3), SVector{3}(1:3)]
+    S12 = compliance[SVector{3}(1:3), SVector{3}(4:6)]
+    S21 = compliance[SVector{3}(4:6), SVector{3}(1:3)]
+    S22 = compliance[SVector{3}(4:6), SVector{3}(4:6)]
+
+    # mass submatrices
+    mass11 = Cab'*mass[SVector{3}(1:3), SVector{3}(1:3)]*Cab
+    mass12 = Cab'*mass[SVector{3}(1:3), SVector{3}(4:6)]*Cab
+    mass21 = Cab'*mass[SVector{3}(4:6), SVector{3}(1:3)]*Cab
[…]

[…]
