### 2022-04-25 byuflowlab/GXBeam.jl#77 @Kevin-Mattheus-Moerman
Apologies for posting a duplicate. Below are the versions I'm using. I'll try to update GXBeam.

```
  "LineSearches"          => v"7.1.1"
  "ColorSchemes"          => v"3.17.1"
  "GXBeam"                => v"0.3.1"
  "IJulia"                => v"1.23.2"
  "Plots"                 => v"1.25.7"
  "PyPlot"                => v"2.10.0"
  "Colors"                => v"0.12.8"
  "Elliptic"              => v"1.0.1"
  "GeometryBasics"        => v"0.4.1"
  "FileIO"                => v"1.13.0"
  "DifferentialEquations" => v"7.1.0"
  "Gridap"                => v"0.17.9"
  "GLMakie"               => v"0.5.2"
```
