### 2022-04-25 #3997 @Kevin-Mattheus-Moerman
@taylormcd @prashjha all issues have now been addressed and I recommend acceptance in JOSS. I wanted to check references, see above, however it seems our bot is not responding to references checking requests this morning. But I am sure you'll check this during the potential acceptance process. Thanks for the opportunity to review this interesting package. All the best.
