+++ src/analyses.jl
@@ -63,7 +63,7 @@ function static_analysis!(system, assembly;
     method=:newton,
     linesearch=LineSearches.BackTracking(maxstep=1e6),
     ftol=1e-9,
-    iterations=1000,
+    iterations=250,
     tvec=0.0,
     show_trace=false,
     reset_state=true)
@@ -226,7 +226,7 @@ function steady_state_analysis!(system, assembly;
     method=:newton,
     linesearch=LineSearches.BackTracking(maxstep=1e6),
     ftol=1e-9,
-    iterations=1000,
+    iterations=250,
     origin=(@SVector zeros(3)),
     linear_velocity=(@SVector zeros(3)),
     angular_velocity=(@SVector zeros(3)),
@@ -235,6 +235,8 @@ function steady_state_analysis!(system, assembly;
     tvec=0.0,
     show_trace=false,
     reset_state=true,
+    expanded=false,
+    structural_damping=false,
     )
 
     # reset state, if specified
@@ -243,13 +245,17 @@ function steady_state_analysis!(system, assembly;
     end
 
     # unpack pre-allocated storage
-    @unpack x, r, K, force_scaling, dynamic_indices = system
+    @unpack x, r, K, force_scaling, dynamic_indices, expanded_indices = system
+
+    # construct expanded state vector (if necessary)
+    if expanded
+        x = get_expanded_state(system, assembly, prescribed_conditions)
+        r = similar(r, length(x))
+        K = similar(K, length(x), length(x))
+    end
 
     # assume converged until proven otherwise
     converged = true
-
-    # structural damping does not impact the solution
-    structural_damping = false
 
     # begin time stepping
     for t in tvec
@@ -272,11 +278,20 @@ function steady_state_analysis!(system, assembly;
         a0 = typeof(linear_acceleration) <: AbstractVector ? SVector{3}(linear_acceleration) : SVector{3}(linear_acceleration(t))
         α0 = typeof(angular_acceleration) <: AbstractVector ? SVector{3}(angular_acceleration) : SVector{3}(angular_acceleration(t))
 
-        f! = (resid, x) -> steady_state_system_residual!(resid, x, dynamic_indices, force_scaling, 
-            structural_damping, assembly, pcond, dload, pmass, gvec, x0, v0, ω0, a0, α0)
+        # residual and jacobian function
+        if expanded
+            f! = (resid, x) -> expanded_system_residual!(resid, x, expanded_indices, force_scaling, 
+                structural_damping, assembly, pcond, dload, pmass, gvec, x0, v0, ω0, a0, α0)
 
-        j! = (jacob, x) -> steady_state_system_jacobian!(jacob, x, dynamic_indices, force_scaling, 
-            structural_damping, assembly, pcond, dload, pmass, gvec, x0, v0, ω0, a0, α0)
+            j! = (jacob, x) -> expanded_system_jacobian!(jacob, x, expanded_indices, force_scaling, 
+                structural_damping, assembly, pcond, dload, pmass, gvec, x0, v0, ω0, a0, α0)
+        else
+            f! = (resid, x) -> steady_state_system_residual!(resid, x, dynamic_indices, force_scaling, 
+                structural_damping, assembly, pcond, dload, pmass, gvec, x0, v0, ω0, a0, α0)
+
+            j! = (jacob, x) -> steady_state_system_jacobian!(jacob, x, dynamic_indices, force_scaling, 
+                structural_damping, assembly, pcond, dload, pmass, gvec, x0, v0, ω0, a0, α0)
+        end
 
         # solve the system of equations
         if linear
@@ -291,7 +306,7 @@ function steady_state_analysis!(system, assembly;
             f!(r, x)
             j!(K, x)
 
-            # update the solution
+            # update the solution               
[…]
