+++ src/element.jl
@@ -82,10 +82,10 @@ matrix system.
 
     icol = icol_elem[ielem]
 
-    CV = SVector(x[icol+12], x[icol+13], x[icol+14])
-    CΩ = SVector(x[icol+15], x[icol+16], x[icol+17])
+    V = SVector(x[icol+12], x[icol+13], x[icol+14])
+    Ω = SVector(x[icol+15], x[icol+16], x[icol+17])
 
-    return CV, CΩ
+    return V, Ω
 end
 
 """
@@ -128,10 +128,10 @@ analysis
     CtCab = C'*Cab
     Qinv = get_Qinv(θ)
 
-    # forces and moments (in the deformed element frame)
+    # forces and moments
     F, M = element_loads(x, ielem, indices.icol_elem, force_scaling)
     
-    # strain and curvature (in the deformed element frame)
+    # strain and curvature
     γ = S11*F + S12*M
     κ = S21*F + S22*M
 
@@ -159,18 +159,7 @@ state analysis
     @unpack L, Cab, C, CtCab, Qinv, mass11, mass12, mass21, mass22, u1, u2, θ1, θ2, 
         u, θ, γ, κ, a, α = properties
 
-    # damping coefficients
-    μ = assembly.elements[ielem].mu
-
-    # damping submatrices
-    μ11 = @SMatrix [μ[1] 0 0; 0 μ[2] 0; 0 0 μ[3]]
-    μ22 = @SMatrix [μ[4] 0 0; 0 μ[5] 0; 0 0 μ[6]]
-
     # rotation parameter matrices
-    C1 = get_C(θ1)
-    C2 = get_C(θ2)
-    Qinv1 = get_Qinv(θ1)
-    Qinv2 = get_Qinv(θ2)
     Q = get_Q(θ)
 
     # linear and angular velocity
@@ -184,58 +173,80 @@ state analysis
     H = CtCab*mass21*CtCab'*V + CtCab*mass22*CtCab'*Ω
 
     # distance from the rotation center
-    Δx1 = assembly.points[assembly.start[ielem]] - x0
-    Δx2 = assembly.points[assembly.stop[ielem]] - x0
     Δx = assembly.elements[ielem].x - x0
 
-    # undeformed linear velocity
-    v1 = v0 + cross(ω0, Δx1 - x0)
-    v2 = v0 + cross(ω0, Δx2 - x0)
+    # rigid body linear and angular velocity
     v = v0 + cross(ω0, Δx)
-
-    # undeformed angular velocity
-    ω1 = ω0
-    ω2 = ω0
     ω = ω0
 
     # linear and angular acceleration
     a += a0 + cross(α0, Δx) + cross(α0, u)
     α += α0
 
-    # linear deflection rates 
-    udot1 = V1 - v1 - cross(ω1, u1)
-    udot2 = V2 - v2 - cross(ω2, u2)
-    udot = (udot1 + udot2)/2
+    # add steady state properties
+    properties = (; properties..., V1, V2, Ω1, Ω2, Q, V, Ω, P, H, γ, κ, v, ω, a, α)
 
-    # angular deflection rates
-    θdot1 = Qinv1*C1*(Ω1 - ω1)
-    θdot2 = Qinv2*C2*(Ω2 - ω2)
-    θdot = (θdot1 + θdot2)/2
-
-    # change in linear and angular displacement
-    Δu = u2 - u1
-    Δθ = θ2 - θ1
-
-    # change in linear and angular displacement rates
-    Δudot = udot2 - udot1
-    Δθdot = θdot2 - θdot1
-
-    # ΔQ matrix (see structural damping theory)
-    ΔQ = get_ΔQ(θ, Δθ, Q)
-
[…]

[…]
