### 2022-05-02 6364954
fix structural damping strain rate
