+++ test/runtests.jl
@@ -313,6 +313,29 @@ end
         @test isapprox(state.points[i].u[3], analytical_deflection(xi), atol=1e-8)
         @test isapprox(state.points[i].theta[2], -4*analytical_slope(xi)/4, atol=1e-7)
     end
+
+    # now check the state variables for a constant mass matrix system
+    system, converged = steady_state_analysis(assembly, 
+        prescribed_conditions = prescribed_conditions,
+        distributed_loads = distributed_loads,
+        expanded = true, 
+        linear = true)
+
+    state = AssemblyState(system, assembly, prescribed_conditions=prescribed_conditions)
+
+    for i = 1:length(assembly.elements)
+        xi = assembly.elements[i].x[1]
+        @test isapprox(state.elements[i].u[3], analytical_deflection(xi), atol=1e-9)
+        @test isapprox(state.elements[i].theta[2], -4*analytical_slope(xi)/4, atol=1e-9)
+        @test isapprox(state.elements[i].Mi[2], -analytical_M(xi), atol=2)
+    end
+
+    for i = 1:length(assembly.points)
+        xi = assembly.points[i][1]
+        @test isapprox(state.points[i].u[3], analytical_deflection(xi), atol=1e-8)
+        @test isapprox(state.points[i].theta[2], -4*analytical_slope(xi)/4, atol=1e-7)
+    end
+
 end
 
 @testset "Linear Analysis of a Beam Under a Linear Distributed Load" begin
@@ -364,6 +387,30 @@ end
     analytical_deflection = (x) -> qmax*(1-x)^2/(120*EI)*(4 - 8*(1-x) + 5*(1-x)^2 - (1-x)^3)
     analytical_slope = (x) -> -qmax*(1-x)/(120*EI)*(8 - 24*(1-x) + 20*(1-x)^2 - 5*(1-x)^3)
     analytical_M = (x) -> qmax/120*(8 - 48*(1-x) + 60*(1-x)^2 - 20*(1-x)^3)
+
+    # test element properties
+    for i = 1:length(assembly.elements)
+        xi = assembly.elements[i].x[1]
+        @test isapprox(state.elements[i].u[3], analytical_deflection(xi), atol=1e-8)
+        @test isapprox(state.elements[i].theta[2], -4*analytical_slope(xi)/4, atol=1e-7)
+        @test isapprox(state.elements[i].Mi[2], -analytical_M(xi), atol=1)
+    end
+
+    # test point properties
+    for i = 1:length(assembly.points)
+        xi = assembly.points[i][1]
+        @test isapprox(state.points[i].u[3], analytical_deflection(xi), atol=1e-8)
+        @test isapprox(state.points[i].theta[2], -4*analytical_slope(xi)/4, atol=1e-8)
+    end
+
+    # now check the state variables for a constant mass matrix system
+    system, converged = steady_state_analysis(assembly, 
+        prescribed_conditions=prescribed_conditions,
+        distributed_loads=distributed_loads, 
+        expanded=true,
+        linear=true)
+
+    state = AssemblyState(system, assembly, prescribed_conditions=prescribed_conditions)
 
     # test element properties
     for i = 1:length(assembly.elements)
@@ -450,6 +497,31 @@ end
         @test isapprox(states[i].points[end].u[3]/L, η_a[i_a], atol=1e-3)
         @test isapprox(states[i].points[end].theta[2], -4*tan(θ_a[i_a]/4), atol=1e-2)
     end
+
+    # perform the same analysis for a constant mass matrix system
[…]
