### 2022-05-02 56f23d4
fix expanded_system_mass_matrix call
