### 2022-05-02 5c55e36
update README to include structural damping
