### 2022-05-03 6507ae9
use correct indices for expanded system during eigenvalue analysis
