+++ src/analyses.jl
@@ -507,11 +507,11 @@ function eigenvalue_analysis!(system, assembly;
         M = similar(K, length(x), length(x))
 
         # solve for the system stiffness matrix
-        expanded_system_jacobian!(K, x, dynamic_indices, force_scaling, structural_damping,
+        expanded_system_jacobian!(K, x, expanded_indices, force_scaling, structural_damping,
             assembly, pcond, dload, pmass, gvec, x0, v0, ω0, a0, α0)
 
         # solve for the system mass matrix
-        expanded_system_mass_matrix!(M, dynamic_indices, force_scaling, assembly, 
+        expanded_system_mass_matrix!(M, expanded_indices, force_scaling, assembly, 
             prescribed_conditions, point_masses)
 
     else
