### 2022-05-13 #3997 @prashjha
Hello @taylormcd, I am reading the draft one last time. If I have any edits or suggestions, I will let you know soon. Meanwhile, could you check the following reference flagged by editorialbot:

Ibrahimbegović, A., & Mikdad, M. A. (1998). Finite rotations in dynamics of beams and implicit time-stepping schemes. International Journal for Numerical Methods in Engineer- ing, 41(5), 781–814. https://doi.org/10.1002/(SICI)1097-0207(19980315)41:5%3C781:: AID-NME308%3E3.0.CO;2-9
