### 2022-05-13 #3997 @prashjha
Hi, @taylormcd, see if these suggestions for the draft makes sense:

- Line 16: missing space between "researchers" and "(Betsch".
- Line 33: "that it is written ..."?
- Line 53: "... accelerations are also provided" is better as you use word "support" in the beginning of this sentence.
- Line 65: "... sectional analysis"?

After you have updated the draft and also fixed the reference I mentiond, can you also do (if not done already) a 'tagged' release of your code, archive the release using zenedo or other methods? Make sure that the title of zenedo archive matches with the title of this JOSS submission.

Once you are done, I will hand your paper to EiC for the final decision.
