+++ paper/paper.bib
@@ -222,15 +222,6 @@ month = jun,
  month = mar,
 }
 
-@article{Ibrahimbegovic1998FiniteRI,
-  title={Finite rotations in dynamics of beams and implicit time-stepping schemes},
-  author={Adnan Ibrahimbegovic and Mazen Al Mikdad},
-  journal={International Journal for Numerical Methods in Engineering},
-  year={1998},
-  volume={41},
-  pages={781-814}
-}
-
 @article{Jelenic1995,
  author = {Jeleni\'c, G. and Saje, M.},
  title = {A kinematically exact space finite strain beam model -- finite element formulation by generalized virtual work principle},
