### 2022-05-17 961099d
fix rotation parameter scaling jacobians
