### 2022-05-17 #3997 @taylormcd
I adopted the first three changes, but not the last one because multiple section analyses may be necessary for a single GXBeam run.  I also tagged a new release and archived it with Zenodo: https://doi.org/10.5281/zenodo.6558458.  Let me know if there's anything else that needs to be done.
