### 2022-05-18 #3997 @prashjha
@taylormcd, compiler is not happy with bib entries (Bezanson2017 and Hodges2006) in paper.bib file. Compiler expects issn value to have exactly 9 elements (e.g., xxxx-xxxx). Could you check these two bib entries?
