### 2022-05-18 #3997 @arfon
Based on past experience, these may in fact be ISBNs rather than ISSN entries.
