+++ paper/paper.bib
@@ -139,7 +139,7 @@ url = {https://doi.org/10.1016/0045-7825(85)90050-7},
  series = {Progress in Astronautics and Aeronautics},
  doi = {10.2514/4.866821},
  url = {https://doi.org/10.2514/4.866821},
- issn = {978-1-56347-697-6},
+ isbn = {978-1-56347-697-6},
 }
 
 @article{Reissner1973,
