### 2022-05-18 #3997 @taylormcd
Bezanson2017 has two ISSN: 0036-1445 and 1095-7200, so I didn't change anything there.  For Hodges, it was a ISBN rather than an ISSN, and the issue is now corrected.
